% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roisplit.R
\name{as.data.frame.roisplit}
\alias{as.data.frame.roisplit}
\title{Convert an `roisplit` object to a data frame}
\usage{
\method{as.data.frame}{roisplit}(x, row.names = NULL, optional = FALSE, data = "analysis", ...)
}
\arguments{
\item{x}{An `roisplit` object.}

\item{row.names}{`NULL` or a character vector giving the row names for the data frame. Missing values are not allowed.}

\item{optional}{A logical: should the column names of the data be checked for legality?}

\item{data}{Either "analysis" or "assessment" to specify which data are returned.}

\item{...}{Additional arguments to be passed to or from methods. Not currently used.}
}
\description{
The analysis or assessment code can be returned as a data
  frame (as dictated by the `data` argument) using
  `as.data.frame.roisplit`. `analysis` and
  `assessment` are shortcuts.
}
