% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{as_roi}
\alias{as_roi}
\title{Convert object to ROI}
\usage{
as_roi(obj, data, ...)
}
\arguments{
\item{obj}{The object to be converted.}

\item{data}{The associated data object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
An ROIVolume or ROISurface object.
}
\description{
Convert the provided object into an ROIVolume or ROISurface object.
}
\examples{
\donttest{
  ds <- gen_sample_dataset(c(5,5,5), 20, blocks=2)
  vox <- sample(which(ds$dataset$mask > 0), 10)
  samp <- data_sample(ds$dataset, vox)
  roi <- as_roi(samp, ds$dataset)
}
}
