% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_result.R
\name{classification_result}
\alias{classification_result}
\alias{binary_classification_result}
\alias{multiway_classification_result}
\alias{regression_result}
\title{Create a \code{classification_result} instance}
\usage{
classification_result(
  observed,
  predicted,
  probs,
  testind = NULL,
  test_design = NULL,
  predictor = NULL
)

binary_classification_result(
  observed,
  predicted,
  probs,
  testind = NULL,
  test_design = NULL,
  predictor = NULL
)

multiway_classification_result(
  observed,
  predicted,
  probs,
  testind = NULL,
  test_design = NULL,
  predictor = NULL
)

regression_result(
  observed,
  predicted,
  testind = NULL,
  test_design = NULL,
  predictor = NULL
)
}
\arguments{
\item{observed}{A vector of observed or true values.}

\item{predicted}{A vector of predicted values.}

\item{probs}{A \code{matrix} of predicted probabilities, with one column per level.}

\item{testind}{The row indices of the test observations (optional).}

\item{test_design}{An optional design for the test data.}

\item{predictor}{An optional predictor object.}
}
\value{
A classification result object, which can be one of: \code{regression_result},
  \code{binary_classification_result}, or \code{multiway_classification_result}.

A binary classification result object, with the class attribute set to "binary_classification_result".
}
\description{
Constructs a classification result object based on the observed and predicted values,
as well as other optional parameters.

Constructs a binary classification result object based on the observed and predicted values,
as well as other optional parameters.
}
\examples{
# A vector of observed values
yobs <- factor(rep(letters[1:4], 5))

# Predicted probabilities
probs <- data.frame(a = runif(1:20), b = runif(1:20), c = runif(1:20), d = runif(1:20))
probs <- sweep(probs, 1, rowSums(probs), "/")

# Get the max probability per row and use this to determine the predicted class
maxcol <- max.col(probs)
predicted <- levels(yobs)[maxcol]

# Construct a classification result
cres <- classification_result(yobs, predicted, probs)

# Compute default performance measures (Accuracy, AUC)
performance(cres)
}
