% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{ClassificationModel}
\alias{ClassificationModel}
\title{create an \code{ClassificationModel} instance}
\usage{
ClassificationModel(caretModel)
}
\arguments{
\item{caretModel}{the underlying caret model object}
}
\description{
create an \code{ClassificationModel} instance
}

