% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_model.R
\name{ClassificationResultSet}
\alias{ClassificationResultSet}
\title{Create an \code{ClassificationResultSet} instance}
\usage{
ClassificationResultSet(blockVar, resultList)
}
\arguments{
\item{blockVar}{the cross-validation blocking variable}

\item{resultList}{a list of type \code{ClassificationResult}}
}
\description{
Create an \code{ClassificationResultSet} instance
}

