% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered_dataset.R
\name{clustered_roi_spec}
\alias{clustered_roi_spec}
\title{Clustered ROI specification (lightweight descriptor)}
\usage{
clustered_roi_spec(seed, neighbors)
}
\value{
A \code{clustered_roi_spec} object.
}
\description{
Clustered ROI specification (lightweight descriptor)
}
\keyword{internal}
