% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional.R
\name{comp_perf}
\alias{comp_perf}
\title{Compile performance results and generate volumetric results}
\usage{
comp_perf(results, region_mask)
}
\arguments{
\item{results}{A list containing the results from regional MVPA analysis.}

\item{region_mask}{A mask representing different regions in the brain image.}
}
\value{
A list containing:
  * vols: A list of volumetric results for each performance metric.
  * perf_mat: A data frame containing the compiled performance results with an added 'roinum' column.
}
\description{
This function compiles the performance results from the regional MVPA analysis
and generates volumetric results based on the region_mask.
}
\keyword{internal}
