% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ConsensusLearner}
\alias{ConsensusLearner}
\title{ConsensusLearner}
\usage{
ConsensusLearner(method = "glmnet", params = list())
}
\arguments{
\item{method}{the consensus method}

\item{params}{additional parameters for the method}
}
\description{
construct specification object for a consensus learner
}

