% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{consensusWeights}
\alias{consensusWeights}
\title{consensusWeights}
\usage{
consensusWeights(x, ...)
}
\arguments{
\item{x}{the result set}

\item{...}{addiitonal args}
}
\description{
compute consensusWeights for a set of classification results
}

