% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossval.R
\name{blocked_cross_validation}
\alias{blocked_cross_validation}
\alias{custom_cross_validation}
\alias{twofold_blocked_cross_validation}
\alias{kfold_cross_validation}
\title{blocked_cross_validation}
\usage{
blocked_cross_validation(block_var, exclude = NULL)

custom_cross_validation(sample_set)

twofold_blocked_cross_validation(block_var, nreps = 10, exclude = NULL)

kfold_cross_validation(len, nfolds = 10, exclude = NULL)
}
\arguments{
\item{block_var}{an integer vector of indicating the cross-validation blocks. Each block is indicating by a unique integer.}

\item{exclude}{an optional vector indicating rows to exclude.}

\item{sample_set}{a \code{list} of training and test sample indices}

\item{len}{the number of observations.}

\item{nfolds}{the number of cross-validation folds.}

\item{exclude}{an optional vector indicating rows to exclude.}
}
\description{
Construct a cross-validation specification using a predefined blocking variable

Construct a cross-validation specification that used a user-supplied set of training and test indices.

Construct a cross-validation specification that randomly partitions input set into two sets of blocks.

Construct a cross-validation specification that randomly partitions input set into \code{nfolds} folds.
}
