% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossval.R
\name{crossv_seq_block}
\alias{crossv_seq_block}
\title{X <- data.frame(x1=rnorm(100), x2=rnorm(100))
y <- rep(letters[1:4], 25)
block_var <- rep(1:4, each=25)
cv <- crossv_seq_block(X,y,2, block_var)}
\usage{
crossv_seq_block(data, y, nfolds, block_var, nreps = 4, block_ind = NULL,
  id = ".id")
}
\description{
X <- data.frame(x1=rnorm(100), x2=rnorm(100))
y <- rep(letters[1:4], 25)
block_var <- rep(1:4, each=25)
cv <- crossv_seq_block(X,y,2, block_var)
}
