% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossval.R
\name{crossv_twofold}
\alias{crossv_twofold}
\title{crossv_twofold}
\usage{
crossv_twofold(data, y, block_var, block_ind, id = ".id", nreps = 15,
  exclude = NULL)
}
\arguments{
\item{data}{the training data}

\item{y}{the response vector}

\item{id}{a character id}

\item{exclude}{optional vector indicating rows to exclude.}
}
\description{
sample dataset via repeated two-fold cross-validation
}
\examples{

X <- data.frame(x1=rnorm(100), x2=rnorm(100))
y <- rep(letters[1:4], 25)
block_var <- rep(1:4, each=20)
cv <- crossv_twofold(X,y,block_var, nreps=10)
}
