% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_model.R
\name{crossval_external}
\alias{crossval_external}
\title{Carry out external cross-validation defined by a \code{FoldIterator} instance and a separate test data set.}
\usage{
crossval_external(crossVal, Y, ROI, Ytest, testROI, subIndices = NULL, model,
  tuneGrid, featureSelector = NULL, parcels = NULL)
}
\arguments{
\item{crossVal}{the \code{CrossValidation} object that defines the splits of training and test samples.}

\item{Y}{the labels or values of dependent variable for the training data.}

\item{ROI}{a region of interest of type \code{ROIVolume} or \code{ROISurface} that encapsulates the training data.}

\item{Ytest}{the test data labels (continuous or factor variables permitted, depending on supplied \code{model}.}

\item{testROI}{a region of interest of type \code{ROIVolume} or \code{ROISurface} that encapsulates the test data.}

\item{subIndices}{an integer vector containg the subset of rows to train model on. if \code{NULL}, train on all rows.}

\item{model}{a \code{caret} model object}

\item{tuneGrid}{model tuning parameters stored in a \code{data.frame}}

\item{featureSelector}{an optional \code{FeatureSelector} instance}

\item{parcels}{an optional parcellation object}

\item{testY}{the labels or values of dependent variable for the test data.}

\item{Xtest}{the test data \code{matrix}}
}
\description{
Carry out external cross-validation defined by a \code{FoldIterator} instance and a separate test data set.
}
\details{
this function should not typically be called by user code.
}

