% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_model.R
\name{crossval_internal}
\alias{crossval_internal}
\title{Carry out internal cross-validation defined by a \code{CrossValidation} instance.}
\usage{
crossval_internal(crossVal, Y, ROI, subIndices, model, tuneGrid,
  featureSelector = NULL, parcels = NULL)
}
\arguments{
\item{crossVal}{the \code{CrossValidation} object that defines the splits of training and test samples.}

\item{Y}{the labels or values of dependent variable.}

\item{ROI}{a region of interest of type \code{ROIVolume} or \code{ROISurface} that encapsulates the training data.}

\item{subIndices}{an integer vector containg the subset of rows to train model on.}

\item{model}{a \code{caret} model object}

\item{tuneGrid}{model tuning parameters stored in a \code{data.frame}}

\item{featureSelector}{an optional \code{FeatureSelector} instance}

\item{parcels}{an optional parcellation object}
}
\description{
Carry out internal cross-validation defined by a \code{CrossValidation} instance.
}
\details{
this function should not typically be called by user code.
}

