% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossval.R
\name{crossval_samples}
\alias{crossval_samples}
\title{crossval_samples}
\usage{
crossval_samples(obj, data, y)
}
\arguments{
\item{obj}{the cross-validation control object}

\item{data}{the data to split up}

\item{y}{the response variable}
}
\description{
crossval_samples
}
