% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcalc.R
\name{cordist}
\alias{cordist}
\alias{mahadist}
\alias{eucdist}
\alias{robustmahadist}
\title{Distance Function Constructors}
\usage{
cordist(labels = NULL, method = c("pearson", "spearman"))

mahadist(labels = NULL)

eucdist(labels = NULL)

robustmahadist(labels = NULL)
}
\arguments{
\item{labels}{A vector of labels associated with the data points.}

\item{method}{The method of distance computation, applicable for `cordist`.}
}
\value{
An object of class `distfun` with a specific method subclass, encapsulating
        all information necessary for computing distances according to the specified method.
}
\description{
These functions provide convenient constructors for various types of distance functions.
Each constructor function initializes a distance function object for use with distance
computation functions, specifying the method and any necessary labels.
}
\details{
The constructors allow for the specification of distance calculation methods and associated labels:
- `cordist` creates a correlation distance function.
- `mahadist` creates a Mahalanobis distance function.
- `eucdist` creates a Euclidean distance function.
- `robustmahadist` creates a robust version of the Mahalanobis distance function.
}
\examples{
dist_obj_1 <- cordist(labels = c("A", "B", "C"), method = "pearson")
dist_obj_2 <- mahadist(labels = c("A", "B", "C"))
dist_obj_3 <- eucdist(labels = c("A", "B", "C"))
dist_obj_4 <- robustmahadist(labels = c("A", "B", "C"))

}
\seealso{
\code{\link{create_dist}} for the underlying constructor used by these functions.
}
\keyword{methods}
