% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchlight.R
\name{do_merge_results}
\alias{do_merge_results}
\title{Merge searchlight results}
\usage{
do_merge_results(r1, good_results)
}
\arguments{
\item{r1}{A list of indices representing the searchlight results to be merged.}

\item{good_results}{A data frame containing the valid searchlight results.}
}
\value{
A combined searchlight result object.
}
\description{
This function merges searchlight results, combining the first result with the rest of the results.
}
\keyword{internal}
