% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchlight.R
\name{do_standard}
\alias{do_standard}
\title{Perform standard searchlight analysis}
\usage{
do_standard(
  model_spec,
  radius,
  mvpa_fun = mvpa_iterate,
  combiner = combine_standard,
  ...,
  drop_probs = FALSE,
  fail_fast = FALSE
)
}
\arguments{
\item{model_spec}{An object specifying the model to be used in the searchlight analysis.}

\item{radius}{The radius of the searchlight sphere.}

\item{mvpa_fun}{The MVPA function to be used in the searchlight analysis (default is \code{mvpa_iterate}).}

\item{combiner}{The function to be used to combine results (default is \code{combine_standard}).}

\item{...}{Additional arguments to be passed to the MVPA function.}
}
\description{
This function performs standard searchlight analysis using a specified model and radius.
It can be customized with different MVPA functions, combiners, and permutation options.
}
\keyword{internal}
