% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remap_rrr_model.R
\name{.fit_residual_rrr}
\alias{.fit_residual_rrr}
\title{Residual reduced-rank regression: (Yw - Xw) ~ Xw}
\usage{
.fit_residual_rrr(Xw, Yw, rank = "auto", max_rank = 20, ridge_lambda = NULL)
}
\value{
A list with fitted residual RRR components.
}
\description{
Residual reduced-rank regression: (Yw - Xw) ~ Xw
}
\keyword{internal}
