% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mvpa_model.R
\name{EnsembleSearchlightModel}
\alias{EnsembleSearchlightModel}
\title{create an \code{EnsembleSearchlightModel} instance}
\usage{
EnsembleSearchlightModel(baseLearners = list(pls = data.frame(ncomp = 1:4)))
}
\arguments{
\item{caretModel}{}
}
\description{
create an \code{EnsembleSearchlightModel} instance
}

