% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_iterate.R
\name{external_crossval}
\alias{external_crossval}
\title{External Cross-Validation}
\usage{
external_crossval(
  roi,
  mspec,
  id,
  compute_performance = TRUE,
  return_fit = FALSE,
  permute = FALSE
)
}
\arguments{
\item{roi}{A list containing train_roi and test_roi elements.}

\item{mspec}{A model specification object.}

\item{id}{A unique identifier for the model.}

\item{compute_performance}{Logical, whether to compute performance metrics (default: TRUE).}

\item{return_fit}{Logical, whether to return the fitted model (default: FALSE).}

\item{permute}{Logical, whether to permute the training labels (default: FALSE).}
}
\value{
A tibble with performance metrics, fitted model (optional), and any warnings or errors.
}
\description{
This function performs external cross-validation on the provided ROI and model specification.
It returns a tibble with performance metrics, fitted model (optional), and any warnings or errors.
}
