% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelection.R
\name{FeatureSelector}
\alias{FeatureSelector}
\title{FeatureSelector
Creates a feature selection specification}
\usage{
FeatureSelector(method, cutoff_type, cutoff_value)
}
\arguments{
\item{method}{the type of feature selection}

\item{cutoff_type}{the type of threshold used to select features.}

\item{cutoff_value}{the numeric vale of the threshold cutoff}
}
\description{
FeatureSelector
Creates a feature selection specification
}
\examples{
fsel <- FeatureSelector("FTest", "top_k", 1000)
fsel <- FeatureSelector("FTest", "top_p", .1)
}

