% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{fit_model.mvpa_model}
\alias{fit_model.mvpa_model}
\title{Fit an MVPA model}
\usage{
\method{fit_model}{mvpa_model}(obj, x, y, wts, param, classProbs, ...)
}
\arguments{
\item{obj}{An object derived from the \code{mvpa_model} class.}

\item{x}{The training data matrix.}

\item{y}{The response vector.}

\item{wts}{Optional class weights (if the underlying model supports it).}

\item{param}{The hyperparameters of the model.}

\item{classProbs}{Logical; if TRUE, class probabilities should be computed (default is FALSE).}

\item{...}{Additional arguments to be passed to the underlying model fitting function.}
}
\value{
A fitted model object with additional attributes "obsLevels" and "problemType".
}
\description{
This function fits a multivariate pattern analysis (MVPA) model to the given data.
}
