% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{FoldIterator}
\alias{FoldIterator}
\title{Create an iterator froma block variable for iterating through the folds during cross-validation}
\usage{
FoldIterator(blockVar)
}
\arguments{
\item{X}{data matrix}

\item{Y}{the labels}

\item{blockVar}{variable denoting the cross-validation folds}
}
\description{
Create an iterator froma block variable for iterating through the folds during cross-validation
}

