% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{format_result}
\alias{format_result}
\title{Format Result Object}
\usage{
format_result(obj, result, error_message, ...)
}
\arguments{
\item{obj}{The result object to be formatted.}

\item{result}{The result object to be formatted.}

\item{error_message}{An optional error message.}

\item{...}{Additional arguments to be passed to the method-specific function.}
}
\description{
Format Result Object
}
\examples{
\donttest{
dataset <- gen_sample_dataset(D = c(6, 6, 6), nobs = 20,
                              response_type = "categorical",
                              data_mode = "image")
cval <- blocked_cross_validation(dataset$design$block_var)
model <- load_model("sda_notune")
mspec <- mvpa_model(model, dataset$dataset, dataset$design,
                    "classification", crossval = cval)
# Typically called internally during processing
format_result(mspec, result = NULL, error_message = "example",
              context = list(test = 1, ytest = factor("a")))
}
}
