% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_rsa_model.R
\name{fut_vector_rsa}
\alias{fut_vector_rsa}
\title{Apply the RSA model in parallel using futures}
\usage{
fut_vector_rsa(mod_spec, sf, method)
}
\arguments{
\item{mod_spec}{The model specification.}

\item{sf}{A tibble containing the data sets and their identifiers.}

\item{method}{Method for computing similarities.}
}
\value{
A combined result of all RSA analyses.
}
\description{
Apply the RSA model in parallel using futures
}
