% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{get_center_ids}
\alias{get_center_ids}
\title{Get Center IDs for Searchlight Iteration}
\usage{
get_center_ids(dataset, ...)
}
\arguments{
\item{dataset}{The dataset object.}

\item{...}{Additional arguments.}
}
\value{
Integer vector of center IDs.
}
\description{
Returns the set of center IDs over which a searchlight analysis iterates.
For volumetric datasets this is the set of nonzero mask voxels; for clustered
datasets it is the sequence of cluster indices.
}
\examples{
ds <- gen_sample_dataset(c(5,5,5), 20)
ids <- get_center_ids(ds$dataset)
length(ids)
}
