% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{get_searchlight}
\alias{get_searchlight}
\title{Generate Searchlight Iterator}
\usage{
get_searchlight(obj, ...)
}
\arguments{
\item{obj}{The input dataset object.}

\item{...}{Additional arguments to methods.}
}
\value{
A searchlight iterator object.
}
\description{
Generate a searchlight iterator suitable for given data.
}
