% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{group_means}
\alias{group_means}
\title{group_means}
\usage{
group_means(X, margin, group)
}
\arguments{
\item{X}{the matrix}

\item{margin}{the margin to average over (1 for rows, 2 for columns)}

\item{group}{the grouping variable, a \code{factor} or \code{integer} vector}
}
\description{
Given a matrix, compute the average vector for each level of a grouping variable.
}
