% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{groupMeans}
\alias{groupMeans}
\title{groupMeans
compute the average vector, splitting a matrix for each level of a grouping variable}
\usage{
groupMeans(X, margin, group)
}
\arguments{
\item{X}{the matrix}

\item{margin}{the margin to average over (1 for rows, 2 for columns)}
}
\description{
groupMeans
compute the average vector, splitting a matrix for each level of a grouping variable
}

