% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{has_crossval}
\alias{has_crossval}
\title{Cross-Validation Availability}
\usage{
has_crossval(obj)
}
\arguments{
\item{obj}{The model object.}
}
\description{
Determine whether cross-validation is specified for the object.
}
\value{
Logical indicating if cross-validation will be performed.
}
\examples{
ds <- gen_sample_dataset(D = c(4, 4, 4), nobs = 10)
cval <- blocked_cross_validation(ds$design$block_var)
mdl <- load_model("sda_notune")
mspec <- mvpa_model(mdl, ds$dataset, ds$design, "classification", crossval = cval)
has_crossval(mspec)
}
}
