% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{has_test_set}
\alias{has_test_set}
\title{Test Set Availability}
\usage{
has_test_set(obj)
}
\arguments{
\item{obj}{Object to query.}
}
\value{
Logical indicating if a test set exists.
}
\description{
Determine whether the object contains a separate test set.
}
\examples{
ds <- gen_sample_dataset(D = c(4, 4, 4), nobs = 10, external_test = TRUE)
has_test_set(ds$design)
}
