% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{has_test_set}
\alias{has_test_set}
\title{Test Set Availability}
\usage{
has_test_set(obj)
}
\arguments{
\item{obj}{The object to check for a test set.}
}
\description{
Check if an object has a test set available.
}
