% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{initialize_configuration}
\alias{initialize_configuration}
\title{initializeROISubset <- function(config) {
  if (!is.null(config$roi_subset)) {
    form <- try(eval(parse(text=config$roi_subset)))
    if (inherits(form, "try-error")) {
      flog.error("could not parse roi_subset parameter: %s", config$roi_subset)
      stop()
    }
    
    if (class(form) != "formula") {
      flog.error("roi_subset argument must be an expression that starts with a ~ character")
      stop()
    }
    
    res <- as.logical(eval(form[[2]], list(x=config$ROIVolume)))
    
    config$ROIVolume[!res] <- 0
    flog.info("roi_subset contains %s voxels", sum(config$ROIVolume > 0))
  }
}}
\usage{
initialize_configuration(args)
}
