% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{initMVPARegional}
\alias{initMVPARegional}
\title{initMVPARegional}
\usage{
initMVPARegional(configFile, args = list(), verbose = FALSE)
}
\arguments{
\item{configFile}{an 'MVPA_Regional' configuration file}

\item{args}{list of addiitonal override arguments}
}
\description{
initMVPARegional
}
