% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_iterate.R
\name{internal_crossval}
\alias{internal_crossval}
\title{Internal Cross-Validation Function}
\usage{
internal_crossval(
  roi,
  mspec,
  id,
  compute_performance = TRUE,
  return_fit = FALSE,
  permute = FALSE
)
}
\arguments{
\item{roi}{A list containing the ROI data.}

\item{mspec}{The model specification.}

\item{id}{The ID of the model.}

\item{compute_performance}{A logical flag indicating whether to compute performance (default: TRUE).}

\item{return_fit}{A logical flag indicating whether to return the fitted model (default: FALSE).}

\item{permute}{A logical flag indicating whether to permute the labels (default: FALSE).}
}
\description{
This is an internal function to perform cross-validation on the given ROI data.
It generates cross-validation samples, fits the model, and computes performance.
}
\keyword{internal}
