% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossval.R
\name{kfold_cross_validation}
\alias{kfold_cross_validation}
\title{kfold_cross_validation}
\usage{
kfold_cross_validation(len, nfolds = 10)
}
\arguments{
\item{len}{the number of observations}

\item{balance}{}
}
\description{
kfold_cross_validation
}
