% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folds.R
\name{KFoldCrossValidation}
\alias{KFoldCrossValidation}
\title{KFoldCrossValidation}
\usage{
KFoldCrossValidation(len, nfolds = 10, balance = FALSE, bootstrap = FALSE)
}
\arguments{
\item{len}{the number of observations}

\item{balance}{}
}
\description{
KFoldCrossValidation
}

