% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{list_model}
\alias{list_model}
\title{a list of model fits}
\usage{
list_model(fits, names = 1:length(fits))
}
\arguments{
\item{fits}{a list of fits}

\item{names}{the names of the fits}
}
\description{
a list of model fits
}
