% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{load_image_data}
\alias{load_image_data}
\title{load_image_data}
\usage{
load_image_data(config, name, indices = NULL)
}
\arguments{
\item{config}{configuration file}

\item{name}{name of file}
}
\description{
load 4D brain data from one of more image provided in \code{name} argument
}
