% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manova_model.R
\name{manova_design}
\alias{manova_design}
\title{manova_design}
\usage{
manova_design(formula, data)
}
\arguments{
\item{formula}{a formula expression specifying MANOVA regression model}

\item{data}{a named list containing the dissimilarity matrices and any other auxiliary variables}
}
\description{
manova_design
}
