% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_iterate.R
\name{manova_iterate}
\alias{manova_iterate}
\title{manova_iterate}
\usage{
manova_iterate(
  mod_spec,
  vox_list,
  ids = 1:length(vox_list),
  batch_size = as.integer(0.1 * length(ids)),
  permute = FALSE
)
}
\arguments{
\item{mod_spec}{a class of type \code{mvpa_model}}

\item{vox_list}{a \code{list} of voxel indices/coordinates}

\item{ids}{a \code{vector} of ids for each voxel set}

\item{batch_size}{The number of voxel sets to process in each batch (defaults to 10% of the total voxel sets).}

\item{permute}{Logical, whether to permute the labels (defaults to FALSE).}
}
\description{
Run MANOVA analysis for each of a list of voxels sets
}
