% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manova_model.R
\name{manova_model}
\alias{manova_model}
\title{Create a MANOVA Model}
\usage{
manova_model(dataset, design)
}
\arguments{
\item{dataset}{An \code{mvpa_dataset} instance.}

\item{design}{A \code{manova_design} instance.}
}
\value{
A MANOVA model object with class attributes "manova_model" and "list".
}
\description{
This function creates a MANOVA model object containing an `mvpa_dataset` instance and a `manova_design` instance.
}
\details{
The function takes an `mvpa_dataset` instance and a `manova_design` instance as input, and returns a MANOVA model object.
The object is a list that contains the dataset and the design with class attributes "manova_model" and "list".
This object can be used for further multivariate statistical analysis using the MANOVA method.
}
\examples{
# Create a MANOVA model using gen_sample_dataset
dset <- gen_sample_dataset(D = c(4, 4, 4), nobs = 24, nlevels = 3, blocks = 3)

# MANOVA design: voxel patterns modelled by condition and block
design <- manova_design(~ Y + block_var, dset$design$train_design)
manova_model_obj <- manova_model(dset$dataset, design)
}
