% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folds.R
\name{MatrixSplitIterator}
\alias{MatrixSplitIterator}
\title{Create an iterator from a matrix \code{X}, a dependent variable \code{Y} and a splitting variable (blockVar)}
\usage{
MatrixSplitIterator(X, Y, blockVar)
}
\arguments{
\item{X}{data matrix}

\item{Y}{the labels}

\item{blockVar}{variable denoting the cross-validation folds}
}
\description{
Create an iterator from a matrix \code{X}, a dependent variable \code{Y} and a splitting variable (blockVar)
}

