% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_analysis.R
\name{matrixToVolumeList}
\alias{matrixToVolumeList}
\title{matrixToVolumeList 
Convenience function to convert a matrix to a list of \code{BrainVolume} instances.}
\usage{
matrixToVolumeList(vox, mat, mask, default = NA)
}
\arguments{
\item{vox}{a matrix of voxel coordinates with same number of rows as \code{mat} argument.}

\item{mat}{a matrix of values where each column is to be converted to \code{BrainVolume} instance.}

\item{mask}{a mask which is used to define the spatial geometry of the output volume}

\item{default}{the default value in the ouput volume. e.g. the value for coordinates not contained in \code{vox}}
}
\description{
matrixToVolumeList 
Convenience function to convert a matrix to a list of \code{BrainVolume} instances.
}

