% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_result.R
\name{multiway_classification_result}
\alias{multiway_classification_result}
\title{create an \code{multiway_classification_result} instance}
\usage{
multiway_classification_result(observed, predicted, probs, test_design = NULL,
  predictor = NULL)
}
\description{
create an \code{multiway_classification_result} instance
}

