% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_analysis.R
\name{mvpa_crossval}
\alias{mvpa_crossval}
\title{mvpa_crossval}
\usage{
mvpa_crossval(dataset, vox, crossVal, model, tuneGrid = NULL,
  featureSelector = NULL, subIndices = NULL)
}
\arguments{
\item{dataset}{an instance of type \code{MVPADataset}}

\item{vox}{a set of voxel coordinates defining the subset of the image dataset to use.}

\item{crossVal}{a cross-validation instance such as \code{BlockedCrossValidation}}

\item{model}{a caret model object}

\item{tuneGrid}{an optional caret-formatted \code{data.frame} containing parameters to be tuned}

\item{featureSelector}{an optional \code{FeatureSelector} instance for selecting relevant features}

\item{subIndices}{an optional vector of row indices (observations) to run cross-validation on.}
}
\description{
cross_validation for an MVPA analysis.
}

