% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{mvpa_searchlight}
\alias{mvpa_searchlight}
\title{mvpa_searchlight}
\usage{
mvpa_searchlight(trainVec, Y, mask, blockVar, radius = 8,
  modelName = "svmLinear", ncores = 2, method = c("randomized",
  "standard"), niter = 4, tuneGrid = NULL, testVec = NULL, testY = NULL)
}
\arguments{
\item{trainVec}{a \code{BrainVector} instance, a 4-dimensional image where the first three dimensons are (x,y,z) and the 4th dimension is the dependent class/variable}

\item{Y}{the dependent variable for training data. If it is a factor, then classification analysis is performed. If it is a continuous variable then regression is performed.
the length of \code{Y} must be the same as the length of the 4th dimension of \code{train_vec}}

\item{mask}{a \code{BrainVolume} instance indicating the inclusion mask for voxels entering the searchlight analysis.}

\item{blockVar}{an \code{integer} vector indicating the blocks to be used for cross-validation. This is usually a variable indicating the scanning "run".
Must be same length as \code{Y}}

\item{radius}{the searchlight radus in mm}

\item{modelName}{the name of the classifcation model to be used}

\item{ncores}{the number of cores for parallel processign (default is 1)}

\item{method}{the type of searchlight (randomized, or standard)}

\item{niter}{the number of searchlight iterations for 'randomized' method}

\item{tuneGrid}{paramter search grid for optimization of classifier tuning parameters}

\item{testVec}{a \code{BrainVector} with the same spatial dimension as shape as \code{trainVec}. If supplied, this data will be held out as a test set.}

\item{testY}{the dependent variable for test data. If supplied, this variable to evaluate classifier model trained on \code{trainVec}.
\code{testY} must be the same as the length of the 4th dimension of \code{test_vec}.}
}
\value{
a named list of \code{BrainVolume} objects, where each name indicates the performance metric and label (e.g. accuracy, AUC)
}
\description{
mvpa_searchlight
}

