% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{MVPADataset}
\alias{MVPADataset}
\title{create an \code{MVPA} instance}
\usage{
MVPADataset(trainVec, Y, mask, blockVar, testVec, testY, modelName = "corsim",
  tuneGrid = NULL)
}
\arguments{
\item{trainVec}{}

\item{Y}{}

\item{mask}{}

\item{blockVar}{}

\item{testVec}{}

\item{testY}{}

\item{modelName}{}

\item{tuneGrid}{}
}
\description{
create an \code{MVPA} instance
}

