% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folds.R
\name{NestedFoldIterator}
\alias{NestedFoldIterator}
\title{NestedFoldIterator}
\usage{
NestedFoldIterator(Y, blockVar, balance = FALSE, bootstrap = FALSE,
  bootstrapMin = 2)
}
\description{
NestedFoldIterator
}

