% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folds.R
\name{NestedMatrixIterator}
\alias{NestedMatrixIterator}
\title{NestedMatrixIterator}
\usage{
NestedMatrixIterator(X, Y, blockVar, balance = FALSE, bootstrap = FALSE,
  bootstrapMin = 2)
}
\arguments{
\item{X}{the data \cdoe{matrix}}

\item{Y}{a vector of labels}
}
\description{
NestedMatrixIterator
}

