% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{nobs}
\alias{nobs}
\title{Get Number of Observations}
\usage{
nobs(x)
}
\arguments{
\item{x}{The input object.}
}
\value{
The number of observations.
}
\description{
Retrieve the number of observations in an object.
}
\examples{
ds <- gen_sample_dataset(c(5,5,5), 20)
nobs(ds$dataset)
}
