% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{nresponses}
\alias{nresponses}
\title{Number of Response Categories}
\usage{
nresponses(x)
}
\arguments{
\item{x}{The object from which to extract the number of categories.}
}
\value{
The number of response categories.
}
\description{
Get the number of response categories or levels.
}
\examples{
des <- mvpa_design(
  train_design = data.frame(y = factor(c("a","b","a","b"))),
  y_train = factor(c("a","b","a","b")),
  block_var = factor(c("1","1","2","2"))
)
nresponses(des)
}
