% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_model.R
\name{NullResult}
\alias{NullResult}
\title{create an \code{ClassificationModel} instance}
\usage{
NullResult(vox, model, observed)
}
\arguments{
\item{caretModel}{the underlying caret model object}
}
\description{
create an \code{ClassificationModel} instance
}

