% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcalc.R
\name{pairwise_dist.cordist}
\alias{pairwise_dist.cordist}
\title{Compute Pairwise Correlation Distances}
\usage{
\method{pairwise_dist}{cordist}(obj, X)
}
\arguments{
\item{X}{Numeric matrix where rows represent observations and columns represent variables.}

\item{dist_obj}{A list containing the method ("correlation") and a block vector to specify
which rows in `X` should not be compared to avoid within-block correlation.}
}
\value{
An object of class `dist` containing the computed correlation distances.
}
\description{
This method computes the pairwise correlation distances for a matrix `X`, excluding
comparisons within the same block as specified by the `dist_obj$block`.
}
\examples{
X <- matrix(rnorm(100), 10, 10)
block <- rep(1:2, each=5)
dist_obj <- list(method = "pearson", block = block)
dist_matrix <- pairwise_dist.correlation(dist_obj, X)

}
