% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcalc.R
\name{pairwise_dist.mahalanobis}
\alias{pairwise_dist.mahalanobis}
\title{Compute Pairwise Mahalanobis Distances}
\usage{
\method{pairwise_dist}{mahalanobis}(obj, X)
}
\arguments{
\item{X}{Numeric matrix where rows represent observations and columns represent variables.}

\item{dist_obj}{A list that might include additional parameters for distance computation, 
currently unused.}
}
\value{
An object of class `dist` containing the computed Mahalanobis distances.
}
\description{
Computes the pairwise Mahalanobis distances using an inverse covariance matrix estimated
from the data matrix `X` with shrinkage.
}
\examples{
X <- matrix(rnorm(100), 10, 10)
dist_matrix <- pairwise_dist.mahalanobis(list(), X)

}
