% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchlight.R
\name{pool_results}
\alias{pool_results}
\title{pool classiifer results collected over a set of overlapping indices}
\usage{
pool_results(good_results)
}
\description{
pool classiifer results collected over a set of overlapping indices
}
