% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{predict.regression_model_fit}
\alias{predict.regression_model_fit}
\title{predict values given new data}
\usage{
\method{predict}{regression_model_fit}(object, newdata, sub_indices = NULL, ...)
}
\arguments{
\item{sub_indices}{a vector of indices used to subset rows of `newdata`}
}
\description{
predict values given new data
}
