% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_rsa_model.R
\name{predict_model.feature_rsa_model}
\alias{predict_model.feature_rsa_model}
\title{Predict Method for Feature RSA Model}
\usage{
predict_model.feature_rsa_model(object, newdata, ...)
}
\arguments{
\item{object}{The trained feature RSA model}

\item{newdata}{New data}

\item{...}{Additional args}
}
\value{
Predicted values in the feature space
}
\description{
Predict Method for Feature RSA Model
}
