% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_rsa_model.R
\name{print.feature_rsa_design}
\alias{print.feature_rsa_design}
\title{Print Method for Feature RSA Design}
\usage{
\method{print}{feature_rsa_design}(x, ...)
}
\arguments{
\item{x}{A feature_rsa_design object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x} (called for side effects).
}
\description{
Print Method for Feature RSA Design
}
\examples{
\donttest{
  D <- dist(matrix(rnorm(5*3), 5, 3))
  labels <- factor(letters[1:5])
  des <- feature_rsa_design(D, labels)
  print(des)
}
}
