% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_sets.R
\name{print.feature_sets}
\alias{print.feature_sets}
\title{Print method for feature_sets}
\usage{
\method{print}{feature_sets}(x, ...)
}
\arguments{
\item{x}{feature_sets object}

\item{...}{ignored}
}
\value{
Invisibly returns the input object \code{x} (called for side effects).
}
\description{
Print method for feature_sets
}
\examples{
X <- matrix(rnorm(10 * 5), 10, 5)
fs <- feature_sets(X, blocks(low = 2, high = 3))
print(fs)
}
