% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_sets_design.R
\name{print.feature_sets_design}
\alias{print.feature_sets_design}
\title{Print method for feature_sets_design}
\usage{
\method{print}{feature_sets_design}(x, ...)
}
\arguments{
\item{x}{A feature_sets_design object}

\item{...}{ignored}
}
\value{
Invisibly returns the input object \code{x} (called for side effects).
}
\description{
Print method for feature_sets_design
}
\examples{
X <- matrix(rnorm(10 * 5), 10, 5)
fs <- feature_sets(X, blocks(low = 2, high = 3))
des <- feature_sets_design(fs, block_var_train = rep(1:2, each = 5))
print(des)
}
