% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_rsa_model.R
\name{print.vector_rsa_model}
\alias{print.vector_rsa_model}
\title{Print Method for vector_rsa_model}
\usage{
\method{print}{vector_rsa_model}(x, ...)
}
\arguments{
\item{x}{An object of class \code{vector_rsa_model}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x} (called for side effects).
}
\description{
Print Method for vector_rsa_model
}
\examples{
ds <- gen_sample_dataset(c(4, 4, 4), nobs = 10, blocks = 2)
D <- as.matrix(dist(matrix(rnorm(5 * 3), 5, 3)))
rownames(D) <- colnames(D) <- letters[1:5]
des <- vector_rsa_design(D, factor(rep(letters[1:5], 2)), rep(1:2, each = 5))
mdl <- vector_rsa_model(ds$dataset, des)
print(mdl)
}
