% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{prob_observed}
\alias{prob_observed}
\title{Probability of Observed Class}
\usage{
prob_observed(x)
}
\arguments{
\item{x}{The object from which to extract the probability.}
}
\value{
A vector of predicted probabilities.
}
\description{
Extract the predicted probability for the observed class.
}
