% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_xdec_model.R
\name{process_roi.naive_xdec_model}
\alias{process_roi.naive_xdec_model}
\title{Naive cross-decoding per ROI}
\usage{
\method{process_roi}{naive_xdec_model}(mod_spec, roi, rnum, ...)
}
\description{
Naive cross-decoding per ROI
}
\keyword{internal}
