% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_result.R
\name{regression_result}
\alias{regression_result}
\title{Create a Regression Result Object}
\usage{
regression_result(
  observed,
  predicted,
  testind = NULL,
  test_design = NULL,
  predictor = NULL
)
}
\arguments{
\item{observed}{A vector of observed values.}

\item{predicted}{A vector of predicted values.}

\item{testind}{A vector of indices for the test data (optional).}

\item{test_design}{The test design (optional).}

\item{predictor}{The predictor used in the regression model (optional).}
}
\value{
A list with class attributes "regression_result", "classification_result", and "list" containing the observed and predicted values, test design, test indices, and predictor.
}
\description{
This function creates a regression result object containing the observed and predicted values, test design, test indices, and predictor.
}
\examples{
res <- regression_result(observed = c(1, 2, 3), predicted = c(1.1, 1.9, 3.2))
performance(res)
}
\seealso{
Other classification_result: 
\code{\link{binary_classification_result}()},
\code{\link{classification_result}()},
\code{\link{multiway_classification_result}()}
}
\concept{classification_result}
