% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMVPA.R
\docType{package}
\name{rMVPA}
\alias{rMVPA}
\alias{rMVPA-package}
\title{rMVPA: A package for multi-voxel pattern analysis (MVPA)}
\description{
The rMVPA package provides tools for running region-of-interest and searchlight MVPA analyses.
}
