% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsa_model.R
\name{rsa_model_mat}
\alias{rsa_model_mat}
\title{Construct a model matrix for an RSA (Representational Similarity Analysis) design}
\usage{
rsa_model_mat(rsa_des)
}
\arguments{
\item{rsa_des}{An RSA design object created by \code{rsa_design()}.}
}
\value{
A named list of vectors, with each vector corresponding to the processed input data of the RSA design.
}
\description{
This function constructs a model matrix for the given RSA design by processing distance matrices and other variables.
}
\details{
The function takes an RSA design object as input and processes the distance matrices and other variables to
construct a model matrix. It handles different types of input matrices, including symmetric and asymmetric
distance matrices, and can include or exclude within-run comparisons based on the RSA design.
}
\examples{
dismat <- dist(matrix(rnorm(100*100), 100, 100))
rdes <- rsa_design(~ dismat, list(dismat=dismat))
rsa_model_mat(rdes)
}
