% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{run_regional}
\alias{run_regional}
\title{Region of interest based MVPA analysis}
\usage{
run_regional(model_spec, region_mask, ...)
}
\arguments{
\item{model_spec}{A \code{mvpa_model} instance containing the model specifications.}

\item{region_mask}{A \code{NeuroVol} or \code{NeuroSurface} object where each region is identified by a unique integer.
Every non-zero set of positive integers will be used to define a set of voxels for classification analysis.}

\item{...}{Extra arguments passed to the specific regional analysis methods.}
}
\value{
A named list of results for each region of interest.
}
\description{
Run a separate MVPA analysis for multiple disjoint regions of interest.
}
\examples{
# TODO: Add an example
}
