% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional.R
\name{run_regional.rsa_model}
\alias{run_regional.rsa_model}
\title{Specialized run_regional method for rsa_model}
\usage{
\method{run_regional}{rsa_model}(
  model_spec,
  region_mask,
  return_fits = FALSE,
  compute_performance = TRUE,
  coalesce_design_vars = FALSE,
  ...
)
}
\arguments{
\item{model_spec}{An \code{rsa_model} object.}

\item{region_mask}{A mask for the brain regions.}

\item{return_fits}{Whether to return each region's fitted model (default \code{FALSE}).}

\item{compute_performance}{\code{logical} indicating whether to compute performance metrics.}

\item{coalesce_design_vars}{If \code{TRUE}, merges design variables into the prediction table (though we typically don't have one for RSA).}

\item{...}{Additional arguments passed to \code{run_regional_base}.}
}
\description{
By default, RSA models do not produce a full prediction table (set \code{return_predictions = FALSE}).
}
