% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional.R
\name{run_regional_base}
\alias{run_regional_base}
\title{A "base" function that does the standard regional analysis pipeline}
\usage{
run_regional_base(
  model_spec,
  region_mask,
  coalesce_design_vars = FALSE,
  processor = NULL,
  verbose = FALSE,
  compute_performance = model_spec$compute_performance,
  return_predictions = model_spec$return_predictions,
  return_fits = model_spec$return_fits,
  ...
)
}
\arguments{
\item{model_spec}{A model specification object.}

\item{region_mask}{A mask representing different brain regions in either a \code{NeuroVol} or \code{NeuroSurface}.}

\item{coalesce_design_vars}{If \code{TRUE}, merges design variables into the prediction table (if present).}

\item{processor}{An optional processor function for each region (ROI).}

\item{verbose}{If \code{TRUE}, prints progress messages.}

\item{compute_performance}{Logical indicating whether to compute performance metrics (defaults to \code{model_spec$compute_performance}).}

\item{return_predictions}{Logical indicating whether to combine a full prediction table (defaults to \code{model_spec$return_predictions}).}

\item{return_fits}{Logical indicating whether to return the fitted models (defaults to \code{model_spec$return_fits}).}

\item{...}{Additional arguments passed along to \code{mvpa_iterate} or other methods.}
}
\value{
A \code{regional_mvpa_result} object containing:
  \itemize{
    \item \code{performance_table}: A \code{tibble} of performance metrics per region
    \item \code{prediction_table}: A \code{tibble} (if \code{return_predictions = TRUE})
    \item \code{vol_results}: A list of volumetric maps (if \code{compute_performance = TRUE})
    \item \code{fits}: A list of model fits if \code{return_fits = TRUE}
    \item \code{model_spec}: The original model specification
  }
}
\description{
A "base" function that does the standard regional analysis pipeline
}
