% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_analysis.R
\name{runAnalysis}
\alias{runAnalysis}
\title{run an MVPA analysis on ablock of data}
\usage{
runAnalysis(object, dataset, ...)
}
\arguments{
\item{object}{the model object}

\item{dataset}{the MVPADataset instance}
}
\value{
a result object
}
\description{
run an MVPA analysis on ablock of data
}

