% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{select_features.catscore}
\alias{select_features.catscore}
\title{Perform feature selection using the CATSCORE method}
\usage{
\method{select_features}{catscore}(obj, X, Y, ranking.score = c("entropy", "avg", "max"), ...)
}
\arguments{
\item{obj}{The feature selection specification created by \code{feature_selector()}.}

\item{X}{The input data matrix.}

\item{Y}{The response variable.}

\item{ranking.score}{The feature score to use. Supported scores are "entropy", "avg", or "max". Default is "entropy".}

\item{...}{Additional arguments (currently unused).}
}
\value{
A logical vector indicating which features to retain.
}
\description{
This function selects features from the input data matrix X using the
CATSCORE method and the provided feature selection specification.
}
\details{
The CATSCORE method computes a correlation adjusted t-test for every column in the matrix using \code{sda.ranking} from the \code{sda} package.
}
\examples{
fsel <- feature_selector("catscore", "top_k", 1000)
X <- as.data.frame(matrix(rnorm(100 * 10), 100, 10))
Y <- rep(letters[1:5], 20)
selected_features <- select_features(fsel, X, Y, ranking.score = "entropy")
}
\seealso{
\code{\link{feature_selector}} for creating a feature selection specification.
}
