% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{select_features.FTest}
\alias{select_features.FTest}
\title{Perform feature selection using the F-test method}
\usage{
\method{select_features}{FTest}(obj, X, Y, ...)
}
\arguments{
\item{obj}{The feature selection specification created by \code{feature_selector()}.}

\item{X}{The input data matrix.}

\item{Y}{The response variable.}
}
\value{
A logical vector indicating which features to retain.
}
\description{
This function selects features from the input data matrix X using the
F-test method and the provided feature selection specification.
}
\details{
The F-test method computes a one-way ANOVA for every column in the feature matrix.
}
\examples{
fsel <- feature_selector("FTest", "top_k", 1000)
X <- as.data.frame(matrix(rnorm(100 * 10), 100, 10))
Y <- rep(letters[1:5], 20)
selected_features <- select_features(fsel, X, Y)
}
\seealso{
\code{\link{feature_selector}} for creating a feature selection specification.
}
