% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{select_features}
\alias{select_features}
\title{Select Features}
\usage{
select_features(obj, X, Y, ...)
}
\arguments{
\item{obj}{The \code{feature_selection} object specifying the feature selection method and its parameters.}

\item{X}{The dataset containing the training features. This can be a matrix or a \code{ROIVolume} or \code{ROISurface} object.}

\item{Y}{The dependent variable as a factor or numeric variable.}

\item{...}{Additional arguments to be passed to the method-specific function.}
}
\value{
A logical vector indicating the columns of \code{X} matrix that were selected.
}
\description{
Given a \code{feature_selection} specification object and a dataset, returns the set of selected features as a binary vector.
}
\examples{
fsel <- feature_selector("FTest", "top_k", 2)
coords <- rbind(c(1,1,1), c(2,2,2), c(3,3,3))
ROI <- neuroim2::ROIVec(neuroim2::NeuroSpace(c(10,10,10)), coords=coords, matrix(rnorm(100*3), 100, 3))
Y <- factor(rep(c("a", "b"), each=50))
featureMask <- select_features(fsel, neuroim2::values(ROI), Y)
sum(featureMask) == 2

fsel2 <- feature_selector("FTest", "top_p", .1)
featureMask <- select_features(fsel2, neuroim2::values(ROI), Y)

}
