% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelection.R
\name{selectFeatures}
\alias{selectFeatures}
\alias{selectFeatures.FTest}
\alias{selectFeatures.catscore}
\title{selectFeatures}
\usage{
selectFeatures(obj, ROI, Y, ...)

\method{selectFeatures}{catscore}(obj, X, Y, ranking.score = c("entropy",
  "avg", "max"))

\method{selectFeatures}{FTest}(obj, ROI, Y)
}
\arguments{
\item{obj}{the \code{FeatureSelection} object}

\item{ROI}{region of interest containing training features, a class of type \code{ROIVolume} or \code{ROISurface}}

\item{Y}{the dependent variable as a \code{factor} or \code{numeric} variable.}

\item{ranking.score}{the feature score: entropy, avg, or max.}

\item{additional}{arguments}
}
\value{
a \code{logical} vector indicating the columns of \code{X} matrix that were selected.
}
\description{
Given a \code{FeatureSelection} specification object and a dataset return the set of selected features as a binary vector.
}
\examples{
fsel <- FeatureSelector("FTest", "top_k", 10)
coords <- rbind(c(1,1,1), c(2,2,2), c(3,3,3))

ROI <- ROIVolume(BrainSpace(c(10,10,10)), coords=coords, matrix(rnorm(100*3), 100, 3))
Y <- factor(rep(c("a", "b"), each=50))
featureMask <- selectFeatures(fsel, ROI, Y)
sum(featureMask) == 10
}

