% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shard_backend.R
\name{shard_cleanup}
\alias{shard_cleanup}
\title{Clean Up Shared-Memory Segments}
\usage{
shard_cleanup(shard_data)
}
\arguments{
\item{shard_data}{List returned by \code{shard_prepare_dataset}
(i.e.\ \code{mod_spec$shard_data}).}
}
\value{
\code{NULL} (invisibly).
}
\description{
Closes shared-memory segments created by \code{\link{use_shard}}.
Called automatically at the end of \code{mvpa_iterate()}; can also be
invoked manually.
}
