% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mvpa_model.R
\name{SimilarityModel}
\alias{SimilarityModel}
\title{create an \code{SimilarityModel} instance}
\usage{
SimilarityModel(type = c("pearson", "spearman", "kendall"))
}
\arguments{
\item{type}{the similarity metric (pearson, kendall, spearman)}
}
\description{
create an \code{SimilarityModel} instance
}

