% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R, R/mvpa_model.R
\name{strip_dataset}
\alias{strip_dataset}
\alias{strip_dataset.default}
\alias{strip_dataset.mvpa_model}
\title{Strip Dataset from Model Specification}
\usage{
strip_dataset(obj, ...)

\method{strip_dataset}{default}(obj, ...)

\method{strip_dataset}{mvpa_model}(obj, ...)
}
\arguments{
\item{obj}{The model specification object.}

\item{...}{Additional arguments.}
}
\value{
The model specification object with the `dataset` element removed or set to NULL.
}
\description{
Removes the potentially large dataset component from a model specification object
to avoid copying it during parallel processing.
}
\examples{
\donttest{
  ds <- gen_sample_dataset(D = c(4, 4, 4), nobs = 20)
  mdl <- load_model("sda_notune")
  mspec <- mvpa_model(mdl, ds$dataset, ds$design, "classification")
  stripped <- strip_dataset(mspec)
  is.null(stripped$dataset)
}
}
