% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_model.R
\name{strip_dataset.default}
\alias{strip_dataset.default}
\title{Default method for strip_dataset generic}
\usage{
\method{strip_dataset}{default}(obj, ...)
}
\arguments{
\item{obj}{The model specification object.}

\item{...}{Additional arguments (ignored).}
}
\value{
The object with `obj$dataset` set to NULL.
}
\description{
Removes the `dataset` element from a model specification object.
}
