% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_result.R
\name{sub_result.multiway_classification_result}
\alias{sub_result.multiway_classification_result}
\title{Subset Multiway Classification Result}
\usage{
\method{sub_result}{multiway_classification_result}(x, indices)
}
\arguments{
\item{x}{An object of class \code{multiway_classification_result} containing the multiway classification results.}

\item{indices}{The set of indices used to subset the results.}
}
\value{
A \code{multiway_classification_result} object containing the subset of results specified by the indices.
}
\description{
This function subsets a multiway classification result based on the provided indices.
}
\seealso{
Other sub_result: 
\code{\link{sub_result.binary_classification_result}()},
\code{\link{sub_result}()}
}
\concept{sub_result}
