% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{sub_result}
\alias{sub_result}
\title{Extract Row-wise Subset of a Result}
\usage{
sub_result(x, indices)
}
\arguments{
\item{x}{The input result object.}

\item{indices}{Row indices to extract.}
}
\value{
A new result object with the specified rows.
}
\description{
Extract a subset of rows from a classification/regression result object.
}
\examples{
cres <- binary_classification_result(
  observed = factor(c("a","b","a")),
  predicted = factor(c("a","b","b")),
  probs = matrix(c(.8,.2,.4,.3,.7,.6), ncol=2, dimnames=list(NULL,c("a","b")))
)
sub_result(cres, 1:2)
}
