% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_rsa_model.R
\name{summary.feature_rsa_model}
\alias{summary.feature_rsa_model}
\title{Summary Method for Feature RSA Model}
\usage{
\method{summary}{feature_rsa_model}(object, ...)
}
\arguments{
\item{object}{The feature RSA model}

\item{...}{Additional args}
}
\value{
A list of summary statistics for the feature RSA model (printed as side effect).
}
\description{
Summary Method for Feature RSA Model
}
\examples{
sample_ds <- gen_sample_dataset(c(4, 4, 4), nobs = 12, blocks = 2)
des <- feature_rsa_design(F = matrix(rnorm(12 * 4), 12, 4),
                          labels = paste0("t", 1:12),
                          block_var = sample_ds$design$block_var)
mdl <- feature_rsa_model(sample_ds$dataset, des, method = "pca")
summary(mdl)
}
