% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R, R/design.R
\name{test_design}
\alias{test_design}
\alias{test_design.mvpa_design}
\title{Test Design Extraction}
\usage{
test_design(obj)

\method{test_design}{mvpa_design}(obj)
}
\arguments{
\item{obj}{The object from which to extract the test design table.}
}
\description{
Return the design table associated with the test set from an object.
}
\examples{
ds <- gen_sample_dataset(D = c(4, 4, 4), nobs = 10, external_test = TRUE)
test_design(ds$design)
}
