% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{test_design}
\alias{test_design}
\title{Test Design Extraction}
\usage{
test_design(obj)
}
\arguments{
\item{obj}{The object from which to extract the test design table.}
}
\description{
Return the design table associated with the test set from an object.
}
