% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_rsa_model.R
\name{train_model.feature_rsa_model}
\alias{train_model.feature_rsa_model}
\title{Train an RSA Model}
\usage{
\method{train_model}{feature_rsa_model}(obj, train_dat, indices, ...)
}
\arguments{
\item{obj}{An object of class \code{feature_rsa_model}}

\item{train_dat}{The training data}

\item{indices}{The indices of the training data}

\item{...}{Additional arguments passed to the training method}
}
\value{
The trained model with updated components
}
\description{
This function trains an RSA (representational similarity analysis) model using the specified method.
}
