% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_model.R
\name{train_model.mvpa_model}
\alias{train_model.mvpa_model}
\title{train_model}
\usage{
\method{train_model}{mvpa_model}(obj, train_dat, y, indices, param = NULL,
  wts = NULL)
}
\arguments{
\item{obj}{an instance of class \code{mvpa_model}}

\item{train_dat}{training data, and instance of class \code{ROIVolume} or \code{ROISurface}}

\item{y}{the dependent variable}

\item{indices}{the spatial indices associated with each column}
}
\description{
train_model
}

