% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_iterate.R
\name{train_model.rsa_model}
\alias{train_model.rsa_model}
\title{Train an RSA Model}
\usage{
\method{train_model}{rsa_model}(
  obj,
  train_dat,
  indices,
  wts = NULL,
  method = c("lm", "rfit", "pearson", "spearman"),
  distmethod = c("pearson", "spearman"),
  ...
)
}
\arguments{
\item{obj}{An object of class \code{rsa_model}.}

\item{train_dat}{The training data.}

\item{indices}{The indices of the training data.}

\item{wts}{Optional, the weights for the model training.}

\item{method}{The method used for model training. One of "lm" (linear regression), "rfit" (robust regression), "pearson" (Pearson correlation), or "spearman" (Spearman correlation). Default is "lm".}

\item{distmethod}{The method used for distance calculation. One of "pearson" (Pearson correlation) or "spearman" (Spearman correlation). Default is "pearson".}

\item{...}{Additional arguments passed to the training method.}
}
\value{
The trained model.
}
\description{
This function trains an RSA (representational similarity analysis) model using the specified method and distance calculation.
}
