% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_rsa_model.R
\name{train_model.vector_rsa_model}
\alias{train_model.vector_rsa_model}
\title{Train a vector RSA model}
\usage{
\method{train_model}{vector_rsa_model}(obj, train_dat, y, indices, ...)
}
\arguments{
\item{obj}{An object of class \code{vector_rsa_model}.}

\item{train_dat}{A data frame or matrix representing the training subset (e.g., voxel intensities).}

\item{y}{Not used in vector RSA (here for consistency with other train_model generics).}

\item{indices}{The spatial indices of the training data (ROI, searchlight, etc.).}

\item{...}{Additional arguments.}
}
\value{
A structure containing "scores" or similar second-order similarity results.
}
\description{
Train a vector RSA model
}
