% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{tune_grid}
\alias{tune_grid}
\title{Extract Tuning Grid}
\usage{
tune_grid(obj, x, y, len)
}
\arguments{
\item{obj}{A model or model specification.}

\item{x}{Training data.}

\item{y}{Response variable.}

\item{len}{Number of parameter sets to generate.}
}
\value{
A data frame of tuning parameter combinations.
}
\description{
Returns the parameter grid used to tune a model.
}
\examples{
ds  <- gen_sample_dataset(D = c(5, 5, 5), nobs = 10)
mdl <- load_model("sda_notune")
tune_grid(mdl, ds$dataset$train_data, ds$design$y_train, len = 1)
}
