% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{tune_model}
\alias{tune_model}
\title{This function finds the best hyperparameters for a given model specification
using a specified tuning grid and cross-validation.}
\usage{
tune_model(mspec, x, y, wts, param, nreps = 10)
}
\arguments{
\item{mspec}{A model specification derived from the \code{mvpa_model} class.}

\item{x}{The training data matrix.}

\item{y}{The response vector.}

\item{wts}{Optional class weights (if the underlying model supports it).}

\item{param}{A \code{data.frame} representing the tuning grid, where
parameter names are indicated by column names.}

\item{nreps}{The number of bootstrap replications (default is 10).}
}
\value{
A data frame containing the best hyperparameter values.
}
\description{
This function finds the best hyperparameters for a given model specification
using a specified tuning grid and cross-validation.
}
\keyword{internal}
