% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_rsa_model.R
\name{vector_rsa_design}
\alias{vector_rsa_design}
\title{Construct a design for a vectorized RSA model}
\usage{
vector_rsa_design(D, labels, block_var)
}
\arguments{
\item{D}{A representational dissimilarity matrix with row.names indicating labels.}

\item{labels}{character vector of labels corresponding to rows in another dataset X.}

\item{block_var}{vector indicating the block (strata) each label belongs to.}
}
\value{
A list containing the elements of the RSA design, class attributes "vector_rsa_design" and "list".
}
\description{
This function constructs a design for an RSA model using a single distance matrix, labels, and blocks.
}
\details{
The function verifies that all labels in the 'labels' are present in the row.names of 'D'.
It also sets up for precomputing cross-block pairs if needed.
}
