% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchlight.R
\name{wrap_out}
\alias{wrap_out}
\title{Wrap output results}
\usage{
wrap_out(perf_mat, dataset, ids = NULL)
}
\arguments{
\item{perf_mat}{A performance matrix containing classifier results.}

\item{dataset}{A dataset object containing the dataset information.}

\item{ids}{An optional vector of voxel IDs.}
}
\value{
A named list of SparseNeuroVec objects representing the wrapped output results.
}
\description{
This function wraps the output results of the performance matrix into a list
of SparseNeuroVec objects for each column in the performance matrix.
}
\keyword{internal}
