% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R, R/design.R, R/mvpa_model.R
\name{y_test}
\alias{y_test}
\alias{y_test.mvpa_design}
\alias{y_test.mvpa_model}
\title{Test Labels/Response Extraction}
\usage{
y_test(obj)

\method{y_test}{mvpa_design}(obj)

\method{y_test}{mvpa_model}(obj)
}
\arguments{
\item{obj}{The object from which to extract the test response variable.}
}
\value{
The test response variable.
}
\description{
Extract the test labels or response variable from an object.
}
\examples{
ds <- gen_sample_dataset(D = c(4, 4, 4), nobs = 10, external_test = TRUE)
y_test(ds$design)
}
