% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{y_train}
\alias{y_train}
\title{Training Labels/Response Extraction}
\usage{
y_train(obj)
}
\arguments{
\item{obj}{The object from which to extract the training response variable.}
}
\description{
Extract the training labels or response variable from an object.
}
