% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{as_roi}
\alias{as_roi}
\title{Convert an object to an ROIVolume or ROISurface object}
\usage{
as_roi(obj, data, ...)
}
\arguments{
\item{obj}{The object to be converted.}

\item{data}{The associated data object used in the conversion process.}

\item{...}{Additional arguments to be passed to the function.}
}
\value{
An ROIVolume or ROISurface object.
}
\description{
The as_roi function is a generic function that converts the provided object
into an ROIVolume or ROISurface object, given the associated data object and
any additional arguments.
}
\keyword{internal}
