% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{as_roi}
\alias{as_roi}
\title{as_roi}
\usage{
as_roi(obj, ...)
}
\arguments{
\item{obj}{the object to convert}

\item{...}{extra args}
}
\description{
convert to an \code{ROIVolume} or \code{ROISurface} object
}
