% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_result.R
\name{binary_classification_result}
\alias{binary_classification_result}
\title{Classification results for binary outcome}
\usage{
binary_classification_result(
  observed,
  predicted,
  probs,
  testind = NULL,
  test_design = NULL,
  predictor = NULL
)
}
\arguments{
\item{observed}{A vector of observed or true values.}

\item{predicted}{A vector of predicted values.}

\item{probs}{A \code{matrix} of predicted probabilities, with one column per level.}

\item{testind}{The row indices of the test observations (optional).}

\item{test_design}{An optional design for the test data.}

\item{predictor}{An optional predictor object.}
}
\value{
A binary classification result object, with the class attribute set to "binary_classification_result".
}
\description{
Constructs a binary classification result object based on the observed and predicted values,
as well as other optional parameters.
}
\examples{
\dontrun{
  obs <- factor(c("a", "b", "a", "b"))
  pred <- factor(c("a", "b", "b", "a"))
  probs <- matrix(c(0.8, 0.2, 0.7, 0.3, 0.4, 0.6, 0.3, 0.7),
                  ncol=2, dimnames=list(NULL, c("a","b")))
  res <- binary_classification_result(obs, pred, probs)
}
}
\seealso{
Other classification_result: 
\code{\link{classification_result}()},
\code{\link{multiway_classification_result}()},
\code{\link{regression_result}()}
}
\concept{classification_result}
