% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_result.R
\name{binary_classification_result}
\alias{binary_classification_result}
\title{binary_classification_result}
\usage{
binary_classification_result(observed, predicted, probs, testDesign,
  predictor = NULL)
}
\arguments{
\item{observed}{the observed classes}

\item{predicted}{the predicted classes}

\item{probs}{the predicted probabilities}
}
\description{
create an \code{binary_classification_result} instance
}

