% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_model.R
\name{binary_classification_result}
\alias{binary_classification_result}
\title{create an \code{binary_classification_result} instance}
\usage{
binary_classification_result(observed, predicted, probs, testDesign,
  predictor = NULL)
}
\description{
create an \code{binary_classification_result} instance
}

