% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_result.R
\name{binary_classification_result}
\alias{binary_classification_result}
\title{binary_classification_result}
\usage{
binary_classification_result(observed, predicted, probs, test_design,
  predictor = NULL)
}
\arguments{
\item{observed}{the observed classes}

\item{predicted}{the predicted classes}

\item{probs}{the predicted probabilities}

\item{test_design}{the design associated with test responses}

\item{predictor}{an object used to predict new responses}
}
\description{
create an \code{binary_classification_result} instance
}
