% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossval.R
\name{blocked_cross_validation}
\alias{blocked_cross_validation}
\title{blocked_cross_validation}
\usage{
blocked_cross_validation(block_var, exclude = NULL)
}
\arguments{
\item{block_var}{an integer vector of indicating the cross-validation blocks. Each block is indicating by a unique integer.}
}
\description{
construct a cross-validation specification using a predefined blocking variable
}
