% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_result.R
\name{classification_result}
\alias{classification_result}
\alias{binary_classification_result}
\alias{multiway_classification_result}
\alias{regression_result}
\title{create a \code{classification_result} instance}
\usage{
classification_result(observed, predicted, probs, test_design = NULL,
  predictor = NULL)

binary_classification_result(observed, predicted, probs, test_design = NULL,
  predictor = NULL)

multiway_classification_result(observed, predicted, probs, test_design = NULL,
  predictor = NULL)

regression_result(observed, predicted, test_design = NULL, predictor = NULL)
}
\arguments{
\item{observed}{the observed values.}

\item{predicted}{the predicted values.}

\item{probs}{a \code{matrix} of predicted probabilities, one column per level.}

\item{test_design}{an optional design for the test data.}

\item{predictor}{an optional predictor object.}
}
\description{
create a \code{classification_result} instance
}
