% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_model.R
\name{ClassificationModel}
\alias{ClassificationModel}
\title{create an \code{ClassificationModel} instance}
\usage{
ClassificationModel(caretModel)
}
\arguments{
\item{caretModel}{the underlying caret model object}
}
\description{
create an \code{ClassificationModel} instance
}

