% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered_dataset.R
\name{clustered_data_sample}
\alias{clustered_data_sample}
\title{Clustered data sample descriptor}
\usage{
clustered_data_sample(seed, neighbors)
}
\description{
Clustered data sample descriptor
}
\keyword{internal}
