% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{compute_performance}
\alias{compute_performance}
\title{compute_performance}
\usage{
compute_performance(obj, result)
}
\arguments{
\item{obj}{the object}

\item{result}{the 'result' to evaluate}
}
\description{
Delegate calculation of performance metrics
}
