% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchlight.R
\name{create_searchlight_performance}
\alias{create_searchlight_performance}
\title{Create a searchlight performance object}
\usage{
create_searchlight_performance(dataset, perf_vec, ids = NULL)
}
\arguments{
\item{dataset}{The dataset object}

\item{perf_vec}{Performance vector for a single metric}

\item{ids}{Optional vector of voxel IDs}
}
\value{
A searchlight_performance object
}
\description{
Create a searchlight performance object
}
\keyword{internal}
