% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossval.R
\name{blocked_cross_validation}
\alias{blocked_cross_validation}
\alias{twofold_blocked_cross_validation}
\alias{kfold_cross_validation}
\title{construct a cross-validation specification using a predefined blocking variable}
\usage{
blocked_cross_validation(block_var, exclude = NULL)

twofold_blocked_cross_validation(block_var, nreps = 10, exclude = NULL)

kfold_cross_validation(len, nfolds = 10, exclude = NULL)
}
\arguments{
\item{block_var}{an integer vector of indicating the cross-validation blocks. Each block is indicating by a unique integer.}

\item{exclude}{an optional vector indicating rows to exclude.}

\item{len}{the number of observations.}

\item{nfolds}{the number of cross-validation folds.}

\item{exclude}{an optional vector indicating rows to exclude.}
}
\description{
construct a cross-validation specification using a predefined blocking variable

construct a cross-validation specification that randomly partitions input set into \code{nfolds} folds.
}
