% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossval.R
\name{bootstrap_blocked_cross_validation}
\alias{bootstrap_blocked_cross_validation}
\alias{blocked_cross_validation}
\alias{custom_cross_validation}
\alias{twofold_blocked_cross_validation}
\alias{kfold_cross_validation}
\title{bootstrap_blocked_cross_validation}
\usage{
bootstrap_blocked_cross_validation(block_var, nreps = 10)

blocked_cross_validation(block_var)

custom_cross_validation(sample_set)

twofold_blocked_cross_validation(block_var, nreps = 10)

kfold_cross_validation(len, nfolds = 10)
}
\arguments{
\item{block_var}{an integer vector of indicating the cross-validation blocks. Each block is indicating by a unique integer.}

\item{nreps}{the number of bootstrap replications}

\item{sample_set}{a \code{list} of training and test sample indices}

\item{len}{the number of observations.}

\item{nfolds}{the number of cross-validation folds.}

\item{block_var}{an integer vector of indicating the cross-validation blocks. Each block is indicating by a unique integer.}
}
\description{
Construct a cross-validation specification using a predefined blocking variable with bootstrap resamples

Construct a cross-validation specification using a predefined blocking variable

Construct a cross-validation specification that used a user-supplied set of training and test indices.

Construct a cross-validation specification that randomly partitions input set into two sets of blocks.

Construct a cross-validation specification that randomly partitions input set into \code{nfolds} folds.
}
