% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossval.R
\name{bootstrap_blocked_cross_validation}
\alias{bootstrap_blocked_cross_validation}
\alias{blocked_cross_validation}
\alias{custom_cross_validation}
\alias{twofold_blocked_cross_validation}
\alias{kfold_cross_validation}
\title{bootstrap_blocked_cross_validation}
\usage{
bootstrap_blocked_cross_validation(block_var, nreps = 10)

blocked_cross_validation(block_var)

custom_cross_validation(sample_set)

twofold_blocked_cross_validation(block_var, nreps = 10)

kfold_cross_validation(len, nfolds = 10)
}
\arguments{
\item{block_var}{an integer vector of indicating the cross-validation blocks. Each block is indicating by a unique integer.}

\item{nreps}{the number of twofold spits}

\item{sample_set}{a \code{list} of training and test sample indices}

\item{len}{the number of observations.}

\item{nfolds}{the number of cross-validation folds.}
}
\description{
Construct a cross-validation specification using a predefined blocking variable with bootstrap resamples

Construct a cross-validation specification using a predefined blocking variable

Construct a cross-validation specification that used a user-supplied set of training and test indices.

Construct a cross-validation specification that randomly partitions input set into two sets of blocks.

Construct a cross-validation specification that randomly partitions input set into \code{nfolds} folds.
}
\examples{

blockvar <- rep(1:5, each=10)
nreps <- 5
cval <- twofold_blocked_cross_validation(blockvar, nreps=nreps)
samples <- crossval_samples(cval, as.data.frame(matrix(rnorm(50*50),50,50)), y=rep(letters[1:5],10))
stopifnot(nrow(samples) == nreps)

cval <- kfold_cross_validation(len=100, nfolds=10)
samples <- crossval_samples(cval, data=as.data.frame(matrix(rnorm(100*10), 100, 10)), y=rep(letters[1:5],20))
stopifnot(nrow(samples) == 10)
}
