% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossval.R
\name{crossv_bootstrap_block}
\alias{crossv_bootstrap_block}
\title{crossv_bootstrap_block}
\usage{
crossv_bootstrap_block(data, y, block_var, nreps = 5, id = ".id")
}
\arguments{
\item{data}{the training data}

\item{y}{the response vector}

\item{block_var}{an \code{integer} \code{vector} defining the cross-validation blocks}

\item{nreps}{the number of resamples}

\item{id}{a character id}
}
\description{
crossv_bootstrap_block
}
\keyword{internal}
