% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossval.R
\name{crossv_k}
\alias{crossv_k}
\title{crossv_k}
\usage{
crossv_k(data, y, k = 5, id = ".id")
}
\arguments{
\item{data}{the training data}

\item{y}{the response vector}

\item{k}{Number of folds (an integer).}

\item{id}{a character id}
}
\description{
sample dataset using k-fold cross-validation
}
