% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossval.R
\name{crossv_twofold}
\alias{crossv_twofold}
\title{crossv_twofold}
\usage{
crossv_twofold(data, y, block_var, block_ind = NULL, id = ".id",
  nreps = 15)
}
\arguments{
\item{data}{the training data}

\item{y}{the response vector}

\item{block_var}{an \code{integer} \code{vector} defining the cross-validation blocks}

\item{block_ind}{the ordered \code{integer} ids of the blocks}

\item{id}{a character id}

\item{nreps}{the number of resamples}
}
\description{
sample dataset via repeated two-fold cross-validation
}
\examples{

X <- data.frame(x1=rnorm(100), x2=rnorm(100))
y <- rep(letters[1:4], 25)
block_var <- rep(1:4, each=25)
cv <- crossv_twofold(X,y,block_var, nreps=10)
}
