% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{crossval_samples}
\alias{crossval_samples}
\title{crossval_samples}
\usage{
crossval_samples(obj, data, y, ...)
}
\arguments{
\item{obj}{A cross-validation control object.}

\item{data}{A data frame containing the predictors.}

\item{y}{A vector containing the response variable.}

\item{...}{Extra arguments passed to the specific cross-validation methods.}
}
\value{
A tibble containing training and testing sets for each fold.
}
\description{
Apply a cross-validation scheme to split the data into training and testing sets.
}
\examples{
cval <- kfold_cross_validation(len = 20, nfolds = 4)
dat  <- as.data.frame(matrix(rnorm(20 * 2), 20, 2))
y    <- factor(rep(letters[1:4], 5))
crossval_samples(cval, dat, y)
}
