% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{custom_performance}
\alias{custom_performance}
\title{custom_performance}
\usage{
custom_performance(x, custom_fun, split_list = NULL)
}
\arguments{
\item{x}{the result object}

\item{custom_fun}{the function used to compute performance metrics, i.e. \code{custom_fun(x)}}

\item{split_list}{the splitting groups}
}
\description{
applies a user-supplied performance metric to a prediction result
}
