% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R, R/resample.R
\name{data_sample}
\alias{data_sample}
\alias{data_sample.mvpa_dataset}
\title{Extract Sample from Dataset}
\usage{
data_sample(obj, vox, ...)

\method{data_sample}{mvpa_dataset}(obj, vox, ...)
}
\arguments{
\item{obj}{The input dataset object.}

\item{vox}{The voxel indices/coordinates.}

\item{...}{Additional arguments to methods.}
}
\value{
A sample extracted from the dataset.
}
\description{
Extract a sample from a given dataset object.
}
