% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{EnsembleSearchlightModel}
\alias{EnsembleSearchlightModel}
\title{create an \code{EnsembleSearchlightModel} instance}
\usage{
EnsembleSearchlightModel(baseLearners = list(pls = data.frame(ncomp = 1:4)))
}
\arguments{
\item{caretModel}{}
}
\description{
create an \code{EnsembleSearchlightModel} instance
}

