% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mvpa_model.R
\name{EnsembleSearchlightModel}
\alias{EnsembleSearchlightModel}
\title{create an \code{EnsembleSearchlightModel} instance}
\usage{
EnsembleSearchlightModel(baseLearners = list(pls = data.frame(ncomp = 1:5),
  lda_thomaz = 1, sda = expand.grid(lambda = c(0.01, 0.2, 0.5, 0.7), diagonal =
  FALSE)))
}
\arguments{
\item{caretModel}{}
}
\description{
create an \code{EnsembleSearchlightModel} instance
}

