% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_rsa_model.R
\name{evaluate_model.feature_rsa_model}
\alias{evaluate_model.feature_rsa_model}
\title{Evaluate Method for Feature RSA Model}
\usage{
evaluate_model.feature_rsa_model(object, predicted, observed, ...)
}
\arguments{
\item{object}{The feature RSA model}

\item{predicted}{Predicted brain activity (X) [nObs x nVox]}

\item{observed}{Observed brain activity (X) [nObs x nVox]}

\item{...}{Additional args}
}
\value{
Performance metrics
}
\description{
Evaluate Method for Feature RSA Model
}
