% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_iterate.R
\name{external_crossval}
\alias{external_crossval}
\title{external_crossval}
\usage{
external_crossval(roi, mspec, id, compute_performance = TRUE,
  return_fit = FALSE)
}
\description{
external_crossval
}
\keyword{internal}
