% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{feature_selector}
\alias{feature_selector}
\title{feature_selector}
\usage{
feature_selector(method, cutoff_type, cutoff_value)
}
\arguments{
\item{method}{the type of feature selection}

\item{cutoff_type}{the type of threshold used to select features.}

\item{cutoff_value}{the numeric vale of the threshold cutoff}
}
\value{
a list with class name equal to the \code{method} arguments
}
\description{
Creates a feature selection specification
}
\details{
The available feature selection methods are:

Ftest: computes a one-way ANOVA for every column in feature matrix
catscore: computes a correlation adjusted t-test for every column in matrix using \code{sda.ranking} from the \code{sda} package.
}
\examples{
fsel <- feature_selector("FTest", "top_k", 1000)
fsel <- feature_selector("FTest", "top_p", .1)
class(fsel) == "FTest"
}
