% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{fit_model}
\alias{fit_model}
\title{Fit Model}
\usage{
fit_model(obj, roi_x, y, wts, param, lev, last, classProbs, ...)
}
\arguments{
\item{obj}{A model fitting object.}

\item{roi_x}{An ROI containing the training data.}

\item{y}{The response vector.}

\item{wts}{A set of case weights.}

\item{param}{Tuning parameters.}

\item{lev}{Unused.}

\item{last}{Unused.}

\item{classProbs}{Unused.}

\item{...}{Additional arguments to be passed to the method-specific function.}
}
\description{
Fit a classification or regression model.
}
