% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folds.R
\name{FoldIterator}
\alias{FoldIterator}
\title{Create an iterator froma block variable for iterating through the folds during cross-validation}
\usage{
FoldIterator(Y, blockVar = NULL, nfolds = 10, balance = FALSE,
  bootstrap = FALSE, bootstrapMin = 2)
}
\arguments{
\item{Y}{the labels}

\item{blockVar}{variable denoting the cross-validation folds}

\item{nfolds}{the number of cross-validation folds: only relvant when blockVar is not supplied}

\item{balance}{try to balance each training sample so that the frequency of labels is equal across groups}

\item{bootstrap}{use bootstrap resampling of the training set}

\item{bootstrapMin}{}
}
\description{
Create an iterator froma block variable for iterating through the folds during cross-validation
}

