% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/folds.R
\name{FoldIterator}
\alias{FoldIterator}
\title{Create an iterator froma block variable for iterating through the folds during cross-validation}
\usage{
FoldIterator(blockVar)
}
\arguments{
\item{blockVar}{variable denoting the cross-validation folds}

\item{X}{data matrix}

\item{Y}{the labels}
}
\description{
Create an iterator froma block variable for iterating through the folds during cross-validation
}

