% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folds.R
\name{FoldIterator}
\alias{FoldIterator}
\title{FoldIterator}
\usage{
FoldIterator(Y, blockVar, balance = FALSE, bootstrap = FALSE,
  bootstrapMin = 2)
}
\arguments{
\item{Y}{the labels}

\item{blockVar}{variable denoting the cross-validation folds}

\item{balance}{try to balance each training sample so that the frequency of labels is equal across groups}

\item{bootstrap}{use bootstrap resampling of the training set}

\item{bootstrapMin}{}
}
\description{
Create an iterator from a block variable for iterating through the folds during cross-validation
}

