% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{format_result}
\alias{format_result}
\title{Format Result Object}
\usage{
format_result(obj, result, error_message, ...)
}
\arguments{
\item{obj}{The result object to be formatted.}

\item{result}{The result object to be formatted.}

\item{error_message}{An optional error message.}

\item{...}{Additional arguments to be passed to the method-specific function.}
}
\description{
Format Result Object
}
