% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{gen_sample_dataset}
\alias{gen_sample_dataset}
\title{gen_sample_dataset}
\usage{
gen_sample_dataset(D, nobs, response_type = c("categorical", "continuous"),
  data_mode = c("image", "surface"), spacing = c(1, 1, 1), blocks = 5,
  nlevels = 5, external_test = FALSE, ntest_obs = nobs)
}
\arguments{
\item{D}{the data dimension(s)}

\item{nobs}{the number of observations}

\item{response_type}{'categorical' or 'continuous'}

\item{data_mode}{'image' or 'surface'}
}
\description{
gen_sample_dataset
}
