% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{gen_sample_dataset}
\alias{gen_sample_dataset}
\title{gen_sample_dataset}
\usage{
gen_sample_dataset(D, nobs, response_type = c("categorical", "continuous"),
  data_mode = c("image", "surface"), spacing = c(1, 1, 1), blocks = 5,
  nlevels = 5, external_test = FALSE, ntest_obs = nobs, split_by = NULL)
}
\arguments{
\item{D}{the data dimension(s)}

\item{nobs}{the number of observations}

\item{response_type}{'categorical' or 'continuous'}

\item{data_mode}{'image' or 'surface'}

\item{spacing}{the voxel spacing}

\item{blocks}{the number of 'blocks' in the data}

\item{nlevels}{the number of category levels}

\item{external_test}{is the test set 'external' to the training set}

\item{ntest_obs}{number of test observations (only relevant if \code{external_test} is true)}

\item{split_by}{compute performance measures for each level of factor}
}
\description{
gen_sample_dataset
}
