% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{get_samples}
\alias{get_samples}
\title{Get Multiple Data Samples}
\usage{
get_samples(obj, vox_list)
}
\arguments{
\item{obj}{The input dataset object.}

\item{vox_list}{A list of vectors containing voxel indices to extract.}
}
\value{
A list of data samples.
}
\description{
Extract multiple data samples based on a list of voxel/index sets from a dataset object.
}
