% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{get_searchlight}
\alias{get_searchlight}
\title{Generate Searchlight Iterator}
\usage{
get_searchlight(obj, ...)
}
\arguments{
\item{obj}{The input dataset object, which should be an instance of a data class compatible with searchlight analysis.}

\item{...}{Additional arguments to be passed to the specific searchlight iterator method for the input object's class.}
}
\value{
A searchlight iterator object that can be used to perform searchlight analysis on the input dataset.
}
\description{
This function generates a searchlight iterator suitable for a given input dataset, such as volumetric or surface data.
}
