% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{get_searchlight}
\alias{get_searchlight}
\title{get_searchlight}
\usage{
get_searchlight(obj, ...)
}
\arguments{
\item{obj}{the object}

\item{...}{extra args}
}
\description{
generate a searchlight iterator appropriate for a given input dataset (e.g. volumetric or surface).
}
