% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossval.R
\name{kfold_cross_validation}
\alias{kfold_cross_validation}
\title{kfold_cross_validation}
\usage{
kfold_cross_validation(len, nfolds = 10)
}
\arguments{
\item{len}{An integer representing the number of observations.}

\item{nfolds}{An integer specifying the number of cross-validation folds.}
}
\value{
An object of class "kfold_cross_validation", "cross_validation", and "list" containing the block_var and nfolds.
}
\description{
Construct a cross-validation specification that randomly partitions the input set into \code{nfolds} folds.
}
\details{
This function creates a k-fold cross-validation scheme for cases where data needs to be split into a specified
number of folds for evaluation. It returns an object of class "kfold_cross_validation", "cross_validation", and "list".
}
\examples{
cval <- kfold_cross_validation(len=100, nfolds=10)
samples <- crossval_samples(cval, data=as.data.frame(matrix(rnorm(100*10), 100, 10)), y=rep(letters[1:5],20))
stopifnot(nrow(samples) == 10)
}
