% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_iterate.R
\name{manova_iterate}
\alias{manova_iterate}
\title{MANOVA Iteration for Voxel Sets}
\usage{
manova_iterate(
  mod_spec,
  vox_list,
  ids = 1:length(vox_list),
  batch_size = as.integer(0.1 * length(ids)),
  permute = FALSE
)
}
\arguments{
\item{mod_spec}{A \code{mvpa_model} object representing the model specification.}

\item{vox_list}{A \code{list} of voxel indices or coordinates.}

\item{ids}{A \code{vector} of IDs for each voxel set.}

\item{batch_size}{An \code{integer} specifying the batch size for processing (default is 10% of the total number of IDs).}

\item{permute}{A \code{logical} indicating whether to permute the voxel sets (default is FALSE).}
}
\value{
A \code{data.frame} containing the MANOVA results for each voxel set.
}
\description{
This function runs a MANOVA analysis for each of a list of voxel sets.
}
