% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_iterate.R
\name{manova_iterate}
\alias{manova_iterate}
\title{manova_iterate}
\usage{
manova_iterate(mod_spec, vox_list, ids = 1:length(vox_list), permute = FALSE)
}
\arguments{
\item{mod_spec}{a class of type \code{mvpa_model}}

\item{vox_list}{a \code{list} of voxel indices/coordinates}

\item{ids}{a \code{vector} of ids for each voxel set}

\item{permute}{permute the labels}
}
\description{
Run MANOVA analysis for each of a list of voxels sets
}
