% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manova_model.R
\name{manova_model}
\alias{manova_model}
\title{manova_model}
\usage{
manova_model(dataset, design)
}
\arguments{
\item{dataset}{a \code{mvpa_dataset} instance}

\item{design}{a \code{manova_design} instance}
}
\description{
manova_model
}
