% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manova_model.R
\name{manova_model}
\alias{manova_model}
\title{Create a MANOVA Model}
\usage{
manova_model(dataset, design)
}
\arguments{
\item{dataset}{An \code{mvpa_dataset} instance.}

\item{design}{A \code{manova_design} instance.}
}
\value{
A MANOVA model object with class attributes "manova_model" and "list".
}
\description{
This function creates a MANOVA model object containing an `mvpa_dataset` instance and a `manova_design` instance.
}
\details{
The function takes an `mvpa_dataset` instance and a `manova_design` instance as input, and returns a MANOVA model object.
The object is a list that contains the dataset and the design with class attributes "manova_model" and "list".
This object can be used for further multivariate statistical analysis using the MANOVA method.
}
\examples{
# Create a MANOVA model
dataset <- create_mvpa_dataset(data_matrix, labels, subject_ids)
formula <- y ~ x1 + x2
data_list <- list(y = dissimilarity_matrix_y, x1 = dissimilarity_matrix_x1, x2 = dissimilarity_matrix_x2)
design <- manova_design(formula, data_list)
manova_model_obj <- manova_model(dataset, design)
}
