% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/folds.R
\name{MatrixFoldIterator}
\alias{MatrixFoldIterator}
\title{Create an iterator from a matrix (X), a dependent variable (Y) and a splitting variable (blockVar)}
\usage{
MatrixFoldIterator(X, Y, blockVar, trainSets = NULL, testSets = NULL)
}
\arguments{
\item{X}{data matrix}

\item{Y}{the labels}

\item{blockVar}{variable denoting the cross-validation folds}
}
\description{
Create an iterator from a matrix (X), a dependent variable (Y) and a splitting variable (blockVar)
}

