% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{MatrixSplitIterator}
\alias{MatrixSplitIterator}
\title{Create an iterator from a matrix (X), a dependent variable (Y) and a splitting variable (blockVar)}
\usage{
MatrixSplitIterator(X, Y, blockVar)
}
\arguments{
\item{X}{data matrix}

\item{Y}{the labels}

\item{blockVar}{variable denoting the cross-validation folds}
}
\description{
Create an iterator from a matrix (X), a dependent variable (Y) and a splitting variable (blockVar)
}

