% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{merge_predictions}
\alias{merge_predictions}
\title{Merge Predictions}
\usage{
merge_predictions(obj1, rest, ...)
}
\arguments{
\item{obj1}{The first object containing predictions.}

\item{rest}{Other objects containing predictions.}

\item{...}{Additional arguments. Methods for this generic may implement specific arguments
such as `weights` to control how predictions are combined.}
}
\value{
A combined object with merged predictions.
}
\description{
Combine predictions from multiple models on the same test set.
}
