% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R, R/model_fit.R
\name{merge_predictions}
\alias{merge_predictions}
\alias{merge_predictions.regression_prediction}
\alias{merge_predictions.classification_prediction}
\title{Merge Predictions}
\usage{
merge_predictions(obj1, rest, ...)

\method{merge_predictions}{regression_prediction}(obj1, rest, ...)

\method{merge_predictions}{classification_prediction}(obj1, rest, ...)
}
\arguments{
\item{obj1}{The first object containing predictions.}

\item{rest}{Other objects containing predictions.}

\item{...}{Additional arguments. Methods for this generic may implement specific arguments
such as `weights` to control how predictions are combined.}
}
\value{
A combined object with merged predictions.
}
\description{
Combine predictions from multiple models on the same test set.
}
\examples{
\donttest{
cres1 <- binary_classification_result(
  observed = factor(c("a","b")),
  predicted = factor(c("a","b")),
  probs = matrix(c(.8,.2,.3,.7), ncol=2, dimnames=list(NULL,c("a","b")))
)
cres2 <- binary_classification_result(
  observed = factor(c("a","b")),
  predicted = factor(c("b","a")),
  probs = matrix(c(.4,.6,.6,.4), ncol=2, dimnames=list(NULL,c("a","b")))
)
merge_predictions(cres1, list(cres2))
}
}
