% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{merge_predictions}
\alias{merge_predictions}
\title{merge_predictions}
\usage{
merge_predictions(obj1, rest, ...)
}
\arguments{
\item{obj1}{the first object}

\item{rest}{the rest of the objects}

\item{...}{extra args}
}
\description{
combine predictions from several models applied to the same test set.
}
