% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{merge_results}
\alias{merge_results}
\title{Merge Multiple Results}
\usage{
merge_results(obj, result_set, indices, id, ...)
}
\arguments{
\item{obj}{The base object containing merge specifications}

\item{result_set}{List of results to be merged}

\item{indices}{List of indices corresponding to each result}

\item{id}{Identifier for the merged result}

\item{...}{Additional arguments passed to specific merge methods}
}
\value{
A merged result object containing:
\itemize{
  \item Combined results from all input objects
  \item Associated indices
  \item Merged metadata
}
}
\description{
Merge Multiple Results
}
