% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{merge_results}
\alias{merge_results}
\title{merge_results}
\usage{
merge_results(x, y, ...)
}
\arguments{
\item{x}{the first result}

\item{y}{the second result}

\item{...}{extra args}
}
\description{
merge two classification/regression results
}
