% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{model_fit}
\alias{model_fit}
\title{the result of a single model fit to a chunk of data}
\usage{
model_fit(model, y, fit, model_type = c("classification", "regression"),
  param, vox_ind, feature_mask = NULL)
}
\arguments{
\item{model}{the caret-style model object}

\item{y}{the predictand}

\item{fit}{the model fit}

\item{model_type}{the problem type: classification or regression}

\item{param}{the model parameters}

\item{vox_ind}{the the voxel indices indicating the data coordinates}

\item{feature_mask}{a logical mask indicating the selected subset of columns}
}
\description{
the result of a single model fit to a chunk of data
}
