% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{MultiWayClassificationResult}
\alias{MultiWayClassificationResult}
\title{create an \code{TwoWayClassification} instance}
\usage{
MultiWayClassificationResult(vox, model, observed, predicted, probs,
  modelFits = NULL)

MultiWayClassificationResult(vox, model, observed, predicted, probs,
  modelFits = NULL)
}
\arguments{
\item{vox}{}

\item{model}{}

\item{observed}{}

\item{predicted}{}

\item{probs}{}

\item{vox}{}

\item{model}{}

\item{observed}{}

\item{predicted}{}

\item{probs}{}
}
\description{
create an \code{TwoWayClassification} instance

create an \code{TwoWayClassification} instance
}

