% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mvpa_model.R
\name{MultiWayClassificationResult}
\alias{MultiWayClassificationResult}
\title{create an \code{TwoWayClassification} instance}
\usage{
MultiWayClassificationResult(observed, predicted, probs, predictor = NULL)
}
\arguments{
\item{observed}{}

\item{predicted}{}

\item{probs}{}
}
\description{
create an \code{TwoWayClassification} instance
}

