% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{MultiWayClassificationResult}
\alias{MultiWayClassificationResult}
\title{create an \code{TwoWayClassification} instance}
\usage{
MultiWayClassificationResult(observed, predicted, probs, predictor = NULL)
}
\arguments{
\item{observed}{}

\item{predicted}{}

\item{probs}{}
}
\description{
create an \code{TwoWayClassification} instance
}

