% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_model.R
\name{MultiWayClassificationResult}
\alias{MultiWayClassificationResult}
\title{create an \code{MultiWayClassification} instance}
\usage{
MultiWayClassificationResult(observed, predicted, probs, testDesign = NULL,
  predictor = NULL)
}
\description{
create an \code{MultiWayClassification} instance
}

