% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_analysis.R
\name{mvpa_crossval}
\alias{mvpa_crossval}
\title{mvpa_crossval}
\usage{
mvpa_crossval(dataset, ROI, crossval, model, tuneGrid = NULL,
  featureSelector = NULL, subIndices = NULL)
}
\arguments{
\item{dataset}{an instance of type \code{MVPADataset}}

\item{ROI}{a class of type \code{ROIVolume} or \code{ROISurface} contianing the training data.}

\item{crossval}{a cross-validation instance of type \code{CrossValidation}}

\item{model}{a \code{caret} model object}

\item{tuneGrid}{an optional caret-formatted \code{data.frame} containing parameters to be tuned}

\item{featureSelector}{an optional \code{FeatureSelector} instance for selecting relevant features}

\item{subIndices}{an optional subset vector of row indices (observations) to run cross-validation on.}
}
\value{
a list consisting of the following named elements:
prediction
predictor  
testIndices 
featureMask
}
\description{
workhorse function for cross_validation in an MVPA analysis
}

