% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{mvpa_dataset}
\alias{mvpa_dataset}
\title{mvpa_dataset}
\usage{
mvpa_dataset(train_data, test_data = NULL, mask, design)
}
\arguments{
\item{train_data}{the training data set: a \code{BrainVector} instance}

\item{test_data}{the test data set: a \code{BrainVector} instance}

\item{mask}{the set of voxels to include: a \code{BrainVolume} instance}
}
\description{
mvpa_dataset
}
