% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{mvpa_dataset}
\alias{mvpa_dataset}
\title{Create an MVPA Dataset Object}
\usage{
mvpa_dataset(train_data, test_data = NULL, mask)
}
\arguments{
\item{train_data}{The training data set: a \code{NeuroVec} instance}

\item{test_data}{Optional test data set: a \code{NeuroVec} instance (default: NULL)}

\item{mask}{The set of voxels to include: a \code{NeuroVol} instance}
}
\value{
An \code{mvpa_dataset} object (S3 class) containing:
  \describe{
    \item{train_data}{The training data as a \code{NeuroVec} instance}
    \item{test_data}{The test data as a \code{NeuroVec} instance (if provided, otherwise NULL)}
    \item{mask}{The binary mask defining valid voxels as a \code{NeuroVol} instance}
  }
}
\description{
Creates a dataset object for MVPA analysis that encapsulates a training dataset, 
an optional test dataset, and a voxel mask.
}
\examples{
# Create dataset from NeuroVec objects
train_vec <- NeuroVec(array(rnorm(1000*100), c(10,10,10,100)))
mask_vol <- NeuroVol(array(1, c(10,10,10)))
dataset <- mvpa_dataset(train_vec, mask=mask_vol)

# Create dataset with test data
test_vec <- NeuroVec(array(rnorm(1000*20), c(10,10,10,20)))
dataset_with_test <- mvpa_dataset(train_vec, test_vec, mask=mask_vol)

}
\seealso{
\code{\link{mvpa_surface_dataset}} for creating surface-based MVPA datasets

\code{\link{mvpa_design}} for creating the corresponding design object
}
