% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{mvpa_design}
\alias{mvpa_design}
\title{mvpa_design}
\usage{
mvpa_design(
  train_design,
  y_train,
  test_design = NULL,
  y_test = NULL,
  block_var = NULL,
  split_by = NULL
)
}
\arguments{
\item{train_design}{A \code{data.frame} containing the training variables.}

\item{y_train}{A \code{formula}, \code{character} name, or \code{factor} specifying the response variable for training.}

\item{test_design}{An optional \code{data.frame} containing the test variables.}

\item{y_test}{An optional \code{formula}, \code{character} name, or \code{factor} specifying the response variable for testing.}

\item{block_var}{An optional \code{formula}, \code{character} name, or \code{integer} vector specifying the blocking variable.}

\item{split_by}{An optional \code{formula} indicating the grouping structure for evaluating test performance.}
}
\description{
Constructs an MVPA design object, defining the classification or regression problem for Multi-Variate Pattern Analysis (MVPA).
}
\examples{

df1 <- data.frame(y = rep(letters[1:4], 5), x1 = rnorm(20), x2 = rnorm(20), block = rep(1:4, each = 5))
des <- mvpa_design(df1, ~ y, block_var = ~ block)

## with a test set

testdes <- data.frame(y = sample(letters[1:4], 10))
des <- mvpa_design(df1, ~ y, block_var = ~ block, test_design = testdes, y_test = ~ y)

}
