% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{mvpa_design}
\alias{mvpa_design}
\title{mvpa_design}
\usage{
mvpa_design(train_design, y_train, test_design = NULL, y_test = NULL,
  block_var = NULL, split_by = NULL)
}
\arguments{
\item{train_design}{a \code{data.frame} containing training variables.}

\item{y_train}{a \code{formula}, \code{character} name or \code{factor} designating the training response.}

\item{test_design}{an optional \code{data.frame} containing test variables.}

\item{y_test}{an optional \code{formula}, \code{character} name or \code{factor} designating the test response.}

\item{block_var}{an optional \code{formula}, \code{character} name or \code{integer} vector designating the block structure.}

\item{split_by}{an optional \code{formula} indicating grouping structure for evaluating test performance.}
}
\description{
mvpa_design
}
\examples{

}
