% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_iterate.R
\name{mvpa_iterate}
\alias{mvpa_iterate}
\title{mvpa_iterate}
\usage{
mvpa_iterate(
  mod_spec,
  vox_list,
  ids = 1:length(vox_list),
  compute_performance = TRUE,
  return_predictions = TRUE,
  return_fits = FALSE,
  batch_size = as.integer(0.1 * length(ids)),
  permute = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{mod_spec}{An object of class \code{mvpa_model} specifying the model.}

\item{vox_list}{A \code{list} of voxel indices or coordinates.}

\item{ids}{A \code{vector} of IDs for each voxel set (defaults to 1:length(vox_list)).}

\item{compute_performance}{Logical, whether to compute and store performance measures for each voxel set (defaults to TRUE).}

\item{return_predictions}{Logical, whether to return row-wise predictions for each voxel set (defaults to TRUE).}

\item{return_fits}{Logical, whether to return the model fit for each voxel set (defaults to FALSE).}

\item{batch_size}{The number of voxel sets to process in each batch (defaults to 10% of the total voxel sets).}

\item{permute}{Logical, whether to permute the labels (defaults to FALSE).}

\item{verbose}{Logical, whether to print progress messages (defaults to TRUE).}
}
\value{
A \code{data.frame} with the results for each voxel set.
}
\description{
Fit a Model for Each Voxel Set in a List Using Parallelization
}
\details{
This function fits a classification or regression model for each voxel set in a list using parallelization.
It can compute and store performance measures for each voxel set, return row-wise predictions, and return the model fit.


This function utilizes parallel processing to speed up the process of fitting the specified model for each voxel set.
The parallelization is achieved using the `furrr` package which provides a parallel backend for the `purrr` package.
By default, it divides the voxel sets into batches and processes them in parallel according to the specified batch size.
The function provides options to control the return of performance measures, predictions, and model fits.
}
