% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_iterate.R
\name{mvpa_iterate}
\alias{mvpa_iterate}
\title{mvpa_iterate}
\usage{
mvpa_iterate(mod_spec, vox_list, ids = 1:length(vox_iter),
  compute_performance = TRUE, return_fits = FALSE)
}
\arguments{
\item{mod_spec}{a class of type \code{mvpa_model}}

\item{vox_list}{a \code{list} of voxel indices/coordinates}

\item{ids}{a \code{vector} of ids for each voxel set}

\item{compute_performance}{compute and store performance measures for each voxel set}

\item{return_fits}{return the model fit for each voxel set?}
}
\description{
Fit a classification/regression model for each voxel set in a list
}
