% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_model.R
\name{mvpa_model}
\alias{mvpa_model}
\title{mvpa_model}
\usage{
mvpa_model(model, dataset, design, model_type = c("classification",
  "regression"), crossval, feature_selector = NULL, tune_grid = NULL,
  performance = NULL, class_metrics = TRUE)
}
\description{
mvpa_model
}

