% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mvpa_analysis.R
\name{mvpa_regional}
\alias{mvpa_regional}
\title{mvpa_regional}
\usage{
mvpa_regional(dataset, regionMask, ncores = 1, savePredictors = FALSE,
  autobalance = FALSE, bootstrap = FALSE, featureSelector = NULL)
}
\arguments{
\item{dataset}{a \code{MVPADataset} instance.}

\item{regionMask}{a \code{BrainVolume} where each region is identified by a unique integer. Every non-zero set of positive integers will be used to define a set of voxels for clasisifcation analysis.}

\item{ncores}{the number of cores for parallel processign (default is 1)}
}
\value{
a named list of \code{BrainVolume} objects, where each name indicates the performance metric and label (e.g. accuracy, AUC)
}
\description{
mvpa_regional
}

