% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{mvpa_regional}
\alias{mvpa_regional}
\title{mvpa_regional}
\usage{
mvpa_regional(bvec, Y, mask, blockVar, modelName = "svmLinear", ncores = 2,
  tuneGrid = NULL)

mvpa_regional(bvec, Y, mask, blockVar, modelName = "svmLinear", ncores = 2,
  tuneGrid = NULL)
}
\arguments{
\item{bvec}{a \code{BrainVector} instance, a 4-dimensional image where the first three dimensions are space (x,y,z) and the 4th dimension is the image/scan/condition.}

\item{Y}{the dependent variable. If it is a factor, then classification analysis is performed. If it is a continuous variable then regression is performed.}

\item{mask}{a \code{BrainVolume} instance indicating the inclusion mask for voxels entering the searchlight analysis.}

\item{blockVar}{an \code{integer} vector indicating the blocks to be used for cross-validation. This is usually a variable indicating the scanning "run".
Must be same length as \code{Y}}

\item{modelName}{the name of the classifcation model to be used}

\item{ncores}{the number of cores for parallel processign (default is 1)}

\item{bvec}{a \code{BrainVector} instance, a 4-dimensional image where the first three dimensions are space (x,y,z) and the 4th dimension is the image/scan/condition.}

\item{Y}{the dependent variable. If it is a factor, then classification analysis is performed. If it is a continuous variable then regression is performed.}

\item{mask}{a \code{BrainVolume} instance indicating the inclusion mask for voxels entering the searchlight analysis.}

\item{blockVar}{an \code{integer} vector indicating the blocks to be used for cross-validation. This is usually a variable indicating the scanning "run".
Must be same length as \code{Y}}

\item{modelName}{the name of the classifcation model to be used}

\item{ncores}{the number of cores for parallel processign (default is 1)}
}
\value{
a named list of \code{BrainVolume} objects, where each name indicates the performance metric and label (e.g. accuracy, AUC)

a named list of \code{BrainVolume} objects, where each name indicates the performance metric and label (e.g. accuracy, AUC)
}
\description{
mvpa_regional

mvpa_regional
}

