% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{mvpa_searchlight}
\alias{mvpa_searchlight}
\title{mvpa_searchlight}
\usage{
mvpa_searchlight(dataset, radius = 8, method = c("randomized", "standard"),
  niter = 4, ncores = 2)
}
\arguments{
\item{dataset}{a \code{MVPADataset} instance.}

\item{radius}{the searchlight radus in mm}

\item{ncores}{the number of cores for parallel processing (default is 1)}

\item{method}{the type of searchlight (randomized, or standard)}

\item{niter}{the number of searchlight iterations for 'randomized' method}
}
\value{
a named list of \code{BrainVolume} objects, where each name indicates the performance metric and label (e.g. accuracy, AUC)
}
\description{
mvpa_searchlight
}

