% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_analysis.R
\name{mvpa_searchlight}
\alias{mvpa_searchlight}
\title{mvpa_searchlight}
\usage{
mvpa_searchlight(dataset, model, crossVal, radius = 8,
  method = c("randomized", "randomized2", "standard"), niter = 4,
  classMetrics = FALSE)
}
\arguments{
\item{dataset}{a \code{MVPADataset} instance.}

\item{radius}{the searchlight radus in mm}

\item{method}{the type of searchlight (randomized, randomized2, or standard)}

\item{niter}{the number of searchlight iterations for 'randomized' method}
}
\value{
a named list of \code{BrainVolume} objects, where each name indicates the performance metric and label (e.g. accuracy, AUC)
}
\description{
mvpa_searchlight
}

