% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{mvpa_surface_dataset}
\alias{mvpa_surface_dataset}
\title{mvpa_surface_dataset}
\usage{
mvpa_surface_dataset(train_data, test_data = NULL, mask, design,
  hemisphere = c("lh", "rh"))
}
\description{
mvpa_surface_dataset
}

