% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{mvpa_surface_dataset}
\alias{mvpa_surface_dataset}
\title{mvpa_surface_dataset}
\usage{
mvpa_surface_dataset(train_data, test_data = NULL, mask = NULL, name = "")
}
\arguments{
\item{train_data}{the training data, must inherit from \code{BrainSurfaceVector} in \code{neurosurf} package.}

\item{test_data}{optional test data, must inherit from \code{BrainSurfaceVector} in \code{neurosurf} package.}

\item{mask}{a binary mask equal to the number of nodes in the training/test data set.}

\item{name}{label to identify the dataset (e.g. "lh" or "rh" to indicate hemisphere)}
}
\description{
Cnstruct a MVPA dataset with a surface-based dataset.
}
