% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_model.R
\name{MVPADataset}
\alias{MVPADataset}
\title{create an \code{MVPA} instance}
\usage{
MVPADataset(trainVec, Y, mask, blockVar, testVec, testY,
  modelName = "corclass", parcellation = NULL, tuneGrid = NULL,
  tuneLength = 1, testSplitVar = NULL, testSplits = NULL,
  trainDesign = NULL, testDesign = NULL)
}
\description{
create an \code{MVPA} instance
}

