% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mvpa_model.R
\name{MVPADataset}
\alias{MVPADataset}
\title{create an \code{MVPA} instance}
\usage{
MVPADataset(trainVec, Y, mask, blockVar, testVec, testY,
  modelName = "corclass", tuneGrid = NULL, tuneLength = 1,
  testSplitVar = NULL, testSplits = NULL)
}
\arguments{
\item{trainVec}{}

\item{Y}{}

\item{mask}{}

\item{blockVar}{}

\item{testVec}{}

\item{testY}{}

\item{modelName}{}

\item{tuneGrid}{}
}
\description{
create an \code{MVPA} instance
}

