% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{MVPADataset}
\alias{MVPADataset}
\title{create an \code{MVPA} instance}
\usage{
MVPADataset(trainVec, Y, mask, blockVar)
}
\arguments{
\item{trainVec}{}

\item{Y}{}

\item{mask}{}

\item{blockVar}{}
}
\description{
create an \code{MVPA} instance
}

