% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folds.R
\name{NestedFoldIterator}
\alias{NestedFoldIterator}
\title{NestedFoldIterator}
\usage{
NestedFoldIterator(Y, block_var, balance = FALSE, bootstrap = FALSE,
  bootstrapMin = 2)
}
\arguments{
\item{Y}{a \code{vector} of labels}

\item{block_var}{an index variable indicating the cross-validation folds.}

\item{balance}{try to balance each training sample so that the frequency of labels is equal across groups}
}
\description{
NestedFoldIterator
}

