% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{nobs}
\alias{nobs}
\title{Get Number of Observations}
\usage{
nobs(x)
}
\arguments{
\item{x}{The input object, which can be of different types, such as a data frame, matrix, or a custom object with a defined `nobs` method.}
}
\value{
The number of observations in the input object. Depending on the object type, this could be the number of rows or elements.
}
\description{
This function retrieves the number of observations from a given object.
}
