% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{nresponses}
\alias{nresponses}
\title{Number of response categories}
\usage{
nresponses(x)
}
\arguments{
\item{x}{The object from which to extract the number of response categories.}
}
\value{
The number of response categories or category levels for the dependent variable.
}
\description{
Get the number of response categories or category levels for the dependent variable.
}
