% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{nresponses}
\alias{nresponses}
\title{Number of Response Categories}
\usage{
nresponses(x)
}
\arguments{
\item{x}{The object from which to extract the number of categories.}
}
\value{
The number of response categories.
}
\description{
Get the number of response categories or levels.
}
\examples{
cres <- binary_classification_result(
  observed = factor(c("a","b")),
  predicted = factor(c("a","b")),
  probs = matrix(c(.8,.2,.3,.7), ncol=2, dimnames=list(NULL,c("a","b")))
)
nresponses(cres)
}
