% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcalc.R
\name{pairwise_dist.euclidean}
\alias{pairwise_dist.euclidean}
\title{Compute Pairwise Euclidean Distances}
\usage{
\method{pairwise_dist}{euclidean}(obj, X)
}
\arguments{
\item{X}{Numeric matrix where rows represent observations and columns represent variables.}

\item{dist_obj}{A list containing possibly additional parameters, currently unused.}
}
\value{
An object of class `dist` containing the computed Euclidean distances.
}
\description{
Computes the pairwise Euclidean distances for a matrix `X`.
}
\examples{
X <- matrix(rnorm(100), 10, 10)
dist_matrix <- pairwise_dist.euclidean(list(), X)

}
