% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcalc.R
\name{pairwise_dist.robustmahadist}
\alias{pairwise_dist.robustmahadist}
\title{Compute Pairwise Robust Mahalanobis Distances}
\usage{
\method{pairwise_dist}{robustmahadist}(obj, X)
}
\arguments{
\item{X}{Numeric matrix where rows represent observations and columns represent variables.}

\item{dist_obj}{A list that might include additional parameters for distance computation, 
currently unused.}
}
\value{
An object of class `dist` containing the computed robust Mahalanobis distances.
}
\description{
Computes the pairwise Mahalanobis distances using a robustly estimated covariance matrix,
which can be more resistant to outliers.
}
\examples{
X <- matrix(rnorm(100), 10, 10)
dist_matrix <- pairwise_dist.robustmahadist(list(), X)

}
