% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R, R/performance.R
\name{performance}
\alias{performance}
\alias{performance.regression_result}
\title{performance}
\usage{
performance(x, ...)

\method{performance}{regression_result}(x, split_list, ...)
}
\arguments{
\item{x}{the result to evaluate performance of}

\item{...}{extra args}

\item{split_list}{split results by indexed sub-groups}
}
\description{
Compute appropriate performance metrics such as accuracy/AUC/RMSE from a classification/regression result
}
