% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{performance.regression_result}
\alias{performance.regression_result}
\title{performance}
\usage{
\method{performance}{regression_result}(x, split_list, ...)
}
\arguments{
\item{x}{the result object}

\item{split_list}{split results by indexed sub-groups.}
}
\description{
performance
}
