% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{performance}
\alias{performance}
\title{performance}
\usage{
performance(x, ...)
}
\arguments{
\item{x}{the result to evaluate performance of}
}
\description{
Calculate performance metrics from a classification/regression result
}

