% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{performance}
\alias{performance}
\title{Compute Performance Metrics for Classification/Regression Results}
\usage{
performance(x, ...)
}
\arguments{
\item{x}{The classification/regression result object to evaluate.}

\item{...}{Additional arguments passed on to the specific performance evaluation method.}
}
\value{
A list of performance metrics for the input classification/regression result object.
}
\description{
This function computes appropriate performance metrics (e.g., accuracy, AUC, RMSE) for a given classification/regression result.
}
