% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{performance}
\alias{performance}
\title{performance}
\usage{
performance(x, ...)
}
\arguments{
\item{x}{the result to evaluate performance of}

\item{...}{extra args}
}
\description{
Compute appropriate performance metrics such as accuracy/AUC/RMSE from a classification/regression result
}
