% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{predict.class_model_fit}
\alias{predict.class_model_fit}
\title{predict}
\usage{
\method{predict}{class_model_fit}(object, newdata, sub_indices = NULL,
  ...)
}
\arguments{
\item{sub_indices}{the subset of row indices to compute predictions on}
}
\description{
predict
}
