% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{predict.regression_model_fit}
\alias{predict.regression_model_fit}
\title{Predict continuous values for a new dataset using a regression model}
\usage{
\method{predict}{regression_model_fit}(object, newdata, sub_indices = NULL, ...)
}
\arguments{
\item{object}{A fitted model object of class \code{regression_model_fit}.}

\item{newdata}{New data to predict on, either as a matrix or a \code{NeuroVec} or \code{NeuroSurfaceVector} object.}

\item{sub_indices}{A vector of indices used to subset rows of `newdata` (optional).}

\item{...}{Additional arguments to be passed to the underlying prediction function.}
}
\value{
A list containing predicted continuous values with class attributes "regression_prediction", "prediction", and "list".
}
\description{
This function predicts continuous values for new data using a fitted regression model.
}
\examples{
# Assuming `fitted_model` is a fitted model object of class `regression_model_fit`
new_data <- iris_dataset$test_data
predictions <- predict(fitted_model, new_data)
}
