% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_rsa_model.R
\name{predict_model.feature_rsa_model}
\alias{predict_model.feature_rsa_model}
\title{Predict Method for Feature RSA Model}
\usage{
predict_model.feature_rsa_model(object, fit, newdata, ...)
}
\arguments{
\item{object}{The trained feature RSA model (method, etc.)}

\item{fit}{The fitted model object from training_model()}

\item{newdata}{New feature data matrix (rows = test obs, cols = features)}

\item{...}{Additional args}
}
\value{
Predicted brain activity \code{X} values (matrix).
}
\description{
Predict Method for Feature RSA Model
}
