% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{predict_model}
\alias{predict_model}
\title{Predict Model Output}
\usage{
predict_model(object, fit, newdata, ...)
}
\arguments{
\item{object}{A fitted model object for which a prediction method is defined.}

\item{fit}{The fitted model object, often returned by `train_model`.
(Note: For some models, `object` itself might be the fit).}

\item{newdata}{New data (e.g., a matrix or data frame) for which to make predictions.
The structure should be compatible with what the model was trained on.}

\item{...}{Additional arguments passed to specific prediction methods.}
}
\value{
Predictions whose structure depends on the specific method (e.g., a vector,
  matrix, or data frame).
}
\description{
Generic function to predict outcomes from a fitted model object using new data.
}
