% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{predicted_class}
\alias{predicted_class}
\title{Calculate the Predicted Class from Probability Matrix}
\usage{
predicted_class(prob)
}
\arguments{
\item{prob}{A matrix of predicted probabilities with column names indicating the classes.}
}
\value{
A vector of predicted classes corresponding to the highest probability for each row in the input matrix.
}
\description{
This function calculates the predicted class from a matrix of predicted probabilities. The class with the highest probability is selected as the predicted class.
}
\examples{
prob <- matrix(c(0.2, 0.8,
                 0.6, 0.4),
               nrow = 2, byrow = TRUE,
               dimnames = list(NULL, c("A", "B")))
predicted_class(prob)
}
