% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{predicted_class}
\alias{predicted_class}
\title{predicted_class}
\usage{
predicted_class(prob)
}
\arguments{
\item{prob}{a matrix of predicted probabilities with column names indicating the classes}
}
\description{
predicted_class
}
