% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_rsa_model.R
\name{print.feature_rsa_model}
\alias{print.feature_rsa_model}
\title{Print Method for Feature RSA Model}
\usage{
\method{print}{feature_rsa_model}(x, ...)
}
\arguments{
\item{x}{A feature_rsa_model object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x} (called for side effects).
}
\description{
Print Method for Feature RSA Model
}
\examples{
\dontrun{
  mdl <- feature_rsa_model(dataset, des)
  print(mdl)
}
}
