% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_rsa_model.R
\name{print.feature_rsa_model}
\alias{print.feature_rsa_model}
\title{Print Method for Feature RSA Model}
\usage{
\method{print}{feature_rsa_model}(x, ...)
}
\arguments{
\item{x}{The feature RSA model}

\item{...}{Additional arguments}
}
\description{
Print Method for Feature RSA Model
}
