% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_sets_design.R
\name{print.feature_sets_design}
\alias{print.feature_sets_design}
\title{Print method for feature_sets_design}
\usage{
\method{print}{feature_sets_design}(x, ...)
}
\arguments{
\item{x}{A feature_sets_design object}

\item{...}{ignored}
}
\value{
Invisibly returns the input object \code{x} (called for side effects).
}
\description{
Print method for feature_sets_design
}
\examples{
\dontrun{
  # print method called on feature_sets_design object
}
}
