% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_rsa_model.R
\name{print.vector_rsa_design}
\alias{print.vector_rsa_design}
\title{Print Method for vector_rsa_design}
\usage{
\method{print}{vector_rsa_design}(x, ...)
}
\arguments{
\item{x}{A vector_rsa_design object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x} (called for side effects).
}
\description{
Print Method for vector_rsa_design
}
\examples{
\donttest{
  D <- as.matrix(dist(matrix(rnorm(5*3), 5, 3)))
  rownames(D) <- colnames(D) <- letters[1:5]
  labels <- factor(rep(letters[1:5], 2))
  block_var <- rep(1:2, each=5)
  des <- vector_rsa_design(D, labels, block_var)
  print(des)
}
}
