% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_rsa_model.R
\name{print.vector_rsa_model}
\alias{print.vector_rsa_model}
\title{Print Method for vector_rsa_model}
\usage{
\method{print}{vector_rsa_model}(x, ...)
}
\arguments{
\item{x}{An object of class \code{vector_rsa_model}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x} (called for side effects).
}
\description{
Print Method for vector_rsa_model
}
\examples{
\donttest{
  S <- matrix(rnorm(10*3), 10, 3)
  labels <- factor(rep(letters[1:5], 2))
  des <- vector_rsa_design(S, labels)
  mdl <- vector_rsa_model(des)
  print(mdl)
}
}
