% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{prob_observed}
\alias{prob_observed}
\title{Probability of Observed Class}
\usage{
prob_observed(x)
}
\arguments{
\item{x}{The object from which to extract the probability.}
}
\value{
A vector of predicted probabilities.
}
\description{
Extract the predicted probability for the observed class.
}
\examples{
cres <- binary_classification_result(
  observed = factor(c("a","b")),
  predicted = factor(c("a","b")),
  probs = matrix(c(.8,.2,.3,.7), ncol=2, dimnames=list(NULL,c("a","b")))
)
prob_observed(cres)
}
