% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_result.R
\name{regression_result}
\alias{regression_result}
\title{create an \code{regression_result} instance}
\usage{
regression_result(observed, predicted, test_design = NULL, predictor = NULL)
}
\arguments{
\item{observed}{the observed values}

\item{predicted}{the predicted values}
}
\description{
create an \code{regression_result} instance
}
