% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_iterate.R
\name{rsa_iterate}
\alias{rsa_iterate}
\title{rsa_iterate}
\usage{
rsa_iterate(mod_spec, vox_list, ids = 1:length(vox_iter),
  regtype = c("rfit", "lm", "pearson", "spearman"),
  distmethod = c("spearman", "pearson"))
}
\arguments{
\item{mod_spec}{a class of type \code{mvpa_model}}

\item{vox_list}{a \code{list} of voxel indices/coordinates}

\item{ids}{a \code{vector} of ids for each voxel set}

\item{regtype}{the analysis method, one of: \code{lm}, \code{rfit}, \code{pearson}, \code{spearman}}

\item{distmethod}{the method used to computer distances between oservations, one of: \code{pearson}, \code{spearman}}
}
\description{
Run RSA analysis for each of a list of voxels sets
}
