% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsa_model.R
\name{rsa_model}
\alias{rsa_model}
\title{rsa_model}
\usage{
rsa_model(Rmat, is_dist = TRUE, dataset, design, crossval, tune_grid = NULL,
  performance = NULL, class_metrics = TRUE)
}
\arguments{
\item{Rmat}{a square relationship matrix with dimensions equal to number of trials in design.}

\item{is_dist}{\code{TRUE} if the \code{Rmat} a distance matrix}

\item{dataset}{a \code{mvpa_dataset} instance}

\item{design}{a \code{rsa_design} instance}

\item{crossval}{a \code{cross_validation} instance}

\item{performance}{an optional custom function for computing performance metrics.}

\item{model}{a estimation model}
}
\description{
rsa_model
}
