% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsa_model.R
\name{rsa_model}
\alias{rsa_model}
\title{rsa_model}
\usage{
rsa_model(dataset, design)
}
\arguments{
\item{dataset}{a \code{mvpa_dataset} instance}

\item{design}{a \code{rsa_design} instance}
}
\description{
rsa_model
}
