% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional.R
\name{run_regional.mvpa_model}
\alias{run_regional.mvpa_model}
\title{Specialized run_regional method for mvpa_model}
\usage{
\method{run_regional}{mvpa_model}(
  model_spec,
  region_mask,
  coalesce_design_vars = FALSE,
  processor = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{model_spec}{An \code{mvpa_model} object.}

\item{region_mask}{A mask for the brain regions.}

\item{coalesce_design_vars}{If \code{TRUE}, merges design variables into the prediction table.}

\item{processor}{An optional processor function for each region.}

\item{verbose}{If \code{TRUE}, prints progress messages.}

\item{...}{Additional arguments passed to \code{run_regional_base}.}
}
\description{
This method wraps \code{run_regional_base} with default arguments for an \code{mvpa_model} object.
}
