% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional.R
\name{run_regional}
\alias{run_regional}
\title{Region of interest based MVPA analysis}
\usage{
run_regional(
  model_spec,
  region_mask,
  return_fits = FALSE,
  compute_performance = TRUE,
  coalesce_design_vars = FALSE
)
}
\arguments{
\item{model_spec}{a \code{mvpa_model} instance}

\item{region_mask}{a \code{NeuroVol} or \code{NeuroSurface} where each region is identified by a unique integer. 
Every non-zero set of positive integers will be used to define a set of voxels for clasisifcation analysis.}

\item{return_fits}{whether to return model fit for every ROI (default is \code{FALSE} to save memory)}

\item{compute_performance}{\code{logical} indicating whether to compute performance measures (e.g. Accuracy, AUC)}
}
\value{
a named list of \code{NeuroVol} objects, where each element contains performance metric and is 
labelled according to the metric used (e.g. Accuracy, AUC)

a \code{list} of type \code{regional_mvpa_result}
}
\description{
Run a separate MVPA analysis for multiple disjoint regions of interest.
}
