% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R, R/regional.R
\name{run_regional}
\alias{run_regional}
\alias{run_regional.mvpa_model}
\alias{run_regional.rsa_model}
\title{Region of interest based MVPA analysis}
\usage{
run_regional(model_spec, region_mask, ...)

\method{run_regional}{mvpa_model}(
  model_spec,
  region_mask,
  return_fits = FALSE,
  return_predictions = TRUE,
  compute_performance = TRUE,
  coalesce_design_vars = FALSE,
  ...
)

\method{run_regional}{rsa_model}(
  model_spec,
  region_mask,
  return_fits = FALSE,
  compute_performance = TRUE,
  regtype = c("pearson", "spearman", "lm", "rfit"),
  distmethod = c("pearson", "spearman"),
  coalesce_design_vars = FALSE,
  ...
)
}
\arguments{
\item{model_spec}{a \code{mvpa_model} instance}

\item{region_mask}{a \code{NeuroVol} or \code{NeuroSurface} where each region is identified by a unique integer. 
Every non-zero set of positive integers will be used to define a set of voxels for classification analysis.}

\item{...}{extra args}

\item{return_fits}{whether to return model fit for every ROI (default is \code{FALSE} to save memory)}

\item{return_predictions}{whether to return full prediction table with per trial probabilities (can be a large table, set \code{FALSE} to limit memory use)}

\item{compute_performance}{\code{logical} indicating whether to compute performance measures (e.g. Accuracy, AUC)}

\item{coalesce_design_vars}{concatenate additional design variables with output stored in `prediction_table`}

\item{regtype}{the regression method ("pearson", "spearman", "lm", or "rfit")}

\item{distmethod}{the distance cmputing method ("pearson" or "spearman")}
}
\value{
a \code{list} of type \code{regional_mvpa_result} containing a named list of \code{NeuroVol} objects, where each element contains performance metric and is 
labelled according to the metric used (e.g. Accuracy, AUC)
}
\description{
Run a separate MVPA analysis for multiple disjoint regions of interest.
}
