% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional.R
\name{run_regional.rsa_model}
\alias{run_regional.rsa_model}
\title{Run regional RSA analysis on a specified RSA model}
\usage{
\method{run_regional}{rsa_model}(
  model_spec,
  region_mask,
  return_fits = FALSE,
  compute_performance = TRUE,
  coalesce_design_vars = FALSE,
  ...
)
}
\arguments{
\item{model_spec}{An object of type \code{rsa_model} specifying the RSA model to be used.}

\item{region_mask}{A mask representing different regions in the brain image.}

\item{return_fits}{Whether to return model fit for every ROI (default is \code{FALSE} to save memory).}

\item{compute_performance}{\code{logical} indicating whether to compute performance measures (e.g. Accuracy, AUC).}

\item{coalesce_design_vars}{Concatenate additional design variables with output stored in `prediction_table`.}

\item{...}{Additional arguments to be passed to the function.}

\item{regtype}{The regression method ("pearson", "spearman", "lm", or "rfit").}

\item{distmethod}{The distance computing method ("pearson" or "spearman").}
}
\value{
A \code{list} of type \code{regional_mvpa_result} containing a named list of \code{NeuroVol} objects,
where each element contains a performance metric and is labeled according to the metric used (e.g. Accuracy, AUC).
}
\description{
This function runs a regional RSA analysis using a specified RSA model and
region mask. The analysis can be customized to return model fits and performance measures.
}
