% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchlight.R
\name{run_searchlight.default}
\alias{run_searchlight.default}
\title{Default method for run_searchlight}
\usage{
\method{run_searchlight}{default}(
  model_spec,
  radius = 8,
  method = c("standard", "randomized", "resampled"),
  niter = 4,
  combiner = "average",
  ...
)
}
\arguments{
\item{model_spec}{The generic model specification object.}

\item{radius}{Numeric searchlight radius (1 to 100).}

\item{method}{Character: "standard" or "randomized".}

\item{niter}{Number of iterations if \code{method="randomized"}.}

\item{combiner}{Either a function that combines partial results or a string
("pool", "average") that selects a built-in combiner.}

\item{...}{Additional arguments passed on to \code{do_standard} or \code{do_randomized}.}
}
\description{
By default, if an object's class does not implement a specific 
\code{run_searchlight.<class>} method, this fallback will call
\code{run_searchlight_base}.
}
