% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchlight.R
\name{run_searchlight}
\alias{run_searchlight}
\title{run_searchlight}
\usage{
run_searchlight(model_spec, radius = 8, method = c("randomized",
  "standard"), niter = 4)
}
\arguments{
\item{model_spec}{a \code{mvpa_model} instance.}

\item{radius}{the searchlight radus in millimeters.}

\item{method}{the type of searchlight (randomized or standard)}

\item{niter}{the number of searchlight iterations (used only for 'randomized' method)}
}
\value{
a named list of \code{BrainVolume} objects, where each element contains a performance metric (e.g. AUC or Accuracy) at every voxel location.
}
\description{
run_searchlight
}

