% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R, R/feature_selection.R
\name{select_features}
\alias{select_features}
\alias{select_features}
\alias{select_features.catscore}
\alias{select_features.FTest}
\title{select_features}
\usage{
select_features(obj, ROI, Y, ...)

select_features(obj, ROI, Y, ...)

\method{select_features}{catscore}(obj, X, Y, ranking.score = c("entropy",
  "avg", "max"))

\method{select_features}{FTest}(obj, X, Y)
}
\arguments{
\item{obj}{the \code{feature_selection} object}

\item{ROI}{region of interest containing training features, a class of type \code{ROIVolume} or \code{ROISurface}}

\item{Y}{the dependent variable as a \code{factor} or \code{numeric} variable.}

\item{ranking.score}{the feature score: entropy, avg, or max.}

\item{additional}{arguments}
}
\value{
a \code{logical} vector indicating the columns of \code{X} matrix that were selected.
}
\description{
performance feature selection on a training ROI.

Given a \code{feature_selection} specification object and a dataset return the set of selected features as a binary vector.
}
\examples{
fsel <- feature_selector("FTest", "top_k", 10)
coords <- rbind(c(1,1,1), c(2,2,2), c(3,3,3))

ROI <- ROIVolume(BrainSpace(c(10,10,10)), coords=coords, matrix(rnorm(100*3), 100, 3))
Y <- factor(rep(c("a", "b"), each=50))
featureMask <- select_features(fsel, ROI, Y)
sum(featureMask) == 10
}
