% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R, R/feature_selection.R
\name{select_features}
\alias{select_features}
\alias{select_features.catscore}
\alias{select_features.FTest}
\title{select_features}
\usage{
select_features(obj, X, Y, ...)

\method{select_features}{catscore}(obj, X, Y, ranking.score = c("entropy",
  "avg", "max"), ...)

\method{select_features}{FTest}(obj, X, Y, ...)
}
\arguments{
\item{obj}{the \code{feature_selection} object}

\item{X}{region of interest containing training features, a class of type \code{ROIVolume} or \code{ROISurface} or \code{matrix}}

\item{Y}{the dependent variable as a \code{factor} or \code{numeric} variable.}

\item{...}{additional arguments}

\item{ranking.score}{the feature score: entropy, avg, or max.}
}
\value{
a \code{logical} vector indicating the columns of \code{X} matrix that were selected.
}
\description{
Given a \code{feature_selection} specification object and a dataset return the set of selected features as a binary vector.
}
\examples{

fsel <- feature_selector("FTest", "top_k", 2)
coords <- rbind(c(1,1,1), c(2,2,2), c(3,3,3))

ROI <- neuroim::ROIVector(neuroim::BrainSpace(c(10,10,10)), coords=coords, matrix(rnorm(100*3), 100, 3))
Y <- factor(rep(c("a", "b"), each=50))
featureMask <- select_features(fsel, ROI@data, Y)
sum(featureMask) == 2

fsel2 <- feature_selector("FTest", "top_p", .1)
featureMask <- select_features(fsel2, ROI@data, Y)

}
