% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelection.R
\name{selectFeatures}
\alias{selectFeatures}
\title{selectFeatures}
\usage{
selectFeatures(obj, X, Y, vox)
}
\arguments{
\item{obj}{the \code{FeatureSelection} object}

\item{X}{the training features as a \code{matrix}}

\item{Y}{the dependent variable as a \code{factor} or \code{numeric} variable.}

\item{vox}{optional coordinates associated with rows of \code{X} matrix.}
}
\value{
a \code{logical} vector indicating the columns of \code{X} matrix that were selected.
}
\description{
Give a \code{FeatureSelection} spec and a dataset return the set of selected features as a binary vector.
}
\examples{
fsel <- FeatureSelector("FTest", "top_k", 1000)
featureMask <- selectFeatures(fsel, X, Y)

}

