% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{SimilarityModel}
\alias{SimilarityModel}
\title{create an \code{SimilarityModel} instance}
\usage{
SimilarityModel(type = c("pearson", "spearman", "kendall"))
}
\arguments{
\item{type}{the similarity metric (pearson, kendall, spearman)}
}
\description{
create an \code{SimilarityModel} instance
}

