% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_result.R
\name{sub_result.multiway_classification_result}
\alias{sub_result.multiway_classification_result}
\alias{sub_result.binary_classification_result}
\title{Subset Multiway Classification Result}
\usage{
\method{sub_result}{multiway_classification_result}(x, indices)

\method{sub_result}{binary_classification_result}(x, indices)
}
\arguments{
\item{x}{An object of class \code{binary_classification_result} containing the binary classification results.}

\item{indices}{The set of indices used to subset the results.}
}
\value{
A \code{multiway_classification_result} object containing the subset of results specified by the indices.

A \code{binary_classification_result} object containing the subset of results specified by the indices.
}
\description{
This function subsets a multiway classification result based on the provided indices.

This function subsets a binary classification result based on the provided indices.
}
\examples{
\dontrun{
  # S3 method - called via sub_result generic
  cres <- multiway_classification_result(
    factor(c("a","b","c")), factor(c("a","b","c")),
    matrix(runif(9), 3, 3, dimnames=list(NULL, c("a","b","c")))
  )
  subset <- sub_result(cres, 1:2)
}
}
