% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R, R/mvpa_result.R
\name{sub_result}
\alias{sub_result}
\alias{sub_result.binary_classification_result}
\title{sub_result}
\usage{
sub_result(x, indices)

\method{sub_result}{binary_classification_result}(x, indices)
}
\arguments{
\item{x}{the result object to subset}

\item{indices}{the set of indices used to subset results}
}
\description{
exctract a row-wise subset of a classification/regression result object.
}
