% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{sub_result}
\alias{sub_result}
\title{Extract Row-wise Subset of a Result}
\usage{
sub_result(x, indices)
}
\arguments{
\item{x}{The input result object.}

\item{indices}{Row indices to extract.}
}
\value{
A new result object with the specified rows.
}
\description{
Extract a subset of rows from a classification/regression result object.
}
