% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{sub_result}
\alias{sub_result}
\title{sub_result}
\usage{
sub_result(x, indices)
}
\arguments{
\item{x}{the result object to subset}

\item{indices}{the row indices}
}
\description{
exctract a row-wise subset of a classification/regression result object.
}
