% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_rsa_model.R
\name{summary.feature_rsa_model}
\alias{summary.feature_rsa_model}
\title{Summary Method for Feature RSA Model}
\usage{
\method{summary}{feature_rsa_model}(object, ...)
}
\arguments{
\item{object}{The feature RSA model}

\item{...}{Additional args}
}
\value{
A list of summary statistics for the feature RSA model (printed as side effect).
}
\description{
Summary Method for Feature RSA Model
}
\examples{
\dontrun{
  mdl <- feature_rsa_model(dataset, des)
  summary(mdl)
}
}
