% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{test_design}
\alias{test_design}
\title{test_design}
\usage{
test_design(obj)
}
\arguments{
\item{obj}{the object to extract test design from}
}
\description{
return design table associated with test set
}
