% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_iterate.R
\name{train_model.manova_model}
\alias{train_model.manova_model}
\title{Train a MANOVA Model}
\usage{
\method{train_model}{manova_model}(obj, train_dat, indices, ...)
}
\arguments{
\item{obj}{An object of class \code{manova_model}.}

\item{train_dat}{The training data.}

\item{indices}{The indices of the training data.}

\item{...}{Additional arguments passed to the training method.}
}
\value{
A named numeric vector of -log(p-values) for each predictor in the MANOVA model.
}
\description{
This function trains a multivariate analysis of variance (MANOVA) model using the specified design.
}
