% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{train_model.mvpa_model}
\alias{train_model.mvpa_model}
\title{Train an MVPA Model}
\usage{
\method{train_model}{mvpa_model}(obj, train_dat, y, indices, wts = NULL, ...)
}
\arguments{
\item{obj}{An object of class \code{mvpa_model}, specifying the MVPA problem.}

\item{train_dat}{Training data, an instance of class \code{ROIVolume} or \code{ROISurface}.}

\item{y}{The dependent variable (response variable), either a numeric vector or a factor.}

\item{indices}{The spatial indices associated with each column.}

\item{wts}{Optional class weights (if the underlying model supports it).}

\item{...}{Additional arguments passed to other methods.}
}
\value{
A model fit object containing the trained model, its fit, the model type (classification or regression), the best tuning parameters, the voxel indices, and the feature mask.
}
\description{
This function trains a Multi-Variate Pattern Analysis (MVPA) model on the provided data, taking care of feature selection, parameter tuning, and model fitting.
}
\examples{
\donttest{
  dset_info <- gen_sample_dataset(D = c(8, 8, 8), nobs = 20,
    response_type = "categorical", data_mode = "image", nlevels = 2)
  cval <- blocked_cross_validation(dset_info$design$block_var)
  sda_model <- load_model("sda_notune")
  mspec <- mvpa_model(
    model = sda_model,
    dataset = dset_info$dataset,
    design = dset_info$design,
    model_type = "classification",
    crossval = cval
  )
  train_model(mspec, dset_info$dataset$train_data,
    dset_info$design$y_train,
    indices = seq_len(ncol(dset_info$dataset$train_data)))
}
}
