% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{train_model}
\alias{train_model}
\title{Train a classification, regression, or representational model.}
\usage{
train_model(obj, ...)
}
\arguments{
\item{obj}{The model specification object.}

\item{...}{Additional arguments to be passed to the method-specific function.}
}
\value{
A trained model object. The exact return value depends on the specific
  method implementation.
}
\description{
This is a generic function that trains a model based on the provided
model specification object. Different model types will have different
methods implemented with specific parameters.
}
