% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{train_model}
\alias{train_model}
\title{Train Model}
\usage{
train_model(obj, ...)
}
\arguments{
\item{obj}{The model specification object.}

\item{...}{Additional arguments to be passed to the method-specific function.}
}
\value{
A trained model object.
}
\description{
Train a classification or regression model.
}
