% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R, R/model_fit.R
\name{train_model}
\alias{train_model}
\alias{train_model.mvpa_model}
\title{train_model}
\usage{
train_model(obj, ...)

\method{train_model}{mvpa_model}(obj, train_dat, y, indices, param = NULL,
  wts = NULL, tune_reps = 10, ...)
}
\arguments{
\item{obj}{the model specification}

\item{...}{extra args}

\item{train_dat}{training data, and instance of class \code{ROIVolume} or \code{ROISurface}}

\item{y}{the dependent variable}

\item{indices}{the spatial indices associated with each column}

\item{param}{optional tuning parameters}

\item{wts}{optional case weights}

\item{tune_reps}{the number of bootstrap replications for parameter tuning (only used when param is not \code{NULL})}
}
\description{
train a classification or regression model

train_model
}
\section{Methods (by class)}{
\itemize{
\item \code{mvpa_model}: train an mvpa_model
}}

