% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_rsa_model.R
\name{train_model.vector_rsa_model}
\alias{train_model.vector_rsa_model}
\title{Train a vector RSA model}
\usage{
\method{train_model}{vector_rsa_model}(obj, train_dat, indices, ...)
}
\arguments{
\item{obj}{An object of class \code{vector_rsa_model}.}

\item{train_dat}{the training data}

\item{indices}{the spatial indices of the training data}

\item{...}{addiitonal arguments}

\item{X}{The data matrix where rows correspond to trials.}
}
\description{
Train a vector RSA model
}
