% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{tune_grid}
\alias{tune_grid}
\title{Tune Grid Extraction}
\usage{
tune_grid(obj, x, y, len)
}
\arguments{
\item{obj}{The model object.}

\item{x}{The training data.}

\item{y}{The response vector.}

\item{len}{The number of elements in the tuning grid.}
}
\description{
Extract the parameter grid to optimize for a model.
}
