% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{tune_grid}
\alias{tune_grid}
\title{tune_grid}
\usage{
tune_grid(obj, x, y, len)
}
\arguments{
\item{obj}{the model object}

\item{x}{the training data}

\item{y}{the response vector}

\item{len}{the number of elements in the tuning grid}
}
\description{
extract the parameter grid to optimize.
}
