% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{tune_model}
\alias{tune_model}
\title{tune_model}
\usage{
tune_model(mspec, x, y, wts, param, nreps = 10)
}
\arguments{
\item{mspec}{the model specification which derives from class \code{mvpa_model}}

\item{x}{the training data matrix}

\item{y}{the response vector}

\item{wts}{optional class weights (if underlying model supports it)}

\item{param}{the tuning grid, should be a \code{data.frame} where parameter names are indicated by column names.}

\item{nreps}{the number of bootstrap replications}
}
\description{
find the best tuning parameters for a model specification
}
