% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{TwoWayClassificationResult}
\alias{TwoWayClassificationResult}
\title{create an \code{TwoWayClassification} instance}
\usage{
TwoWayClassificationResult(vox, model, observed, predicted, probs,
  finalFit = NULL)
}
\arguments{
\item{vox}{}

\item{model}{}

\item{observed}{}

\item{predicted}{}

\item{probs}{}

\item{finalFit}{}
}
\description{
create an \code{TwoWayClassification} instance
}

