% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpa_model.R
\name{TwoWayClassificationResult}
\alias{TwoWayClassificationResult}
\title{create an \code{TwoWayClassification} instance}
\usage{
TwoWayClassificationResult(observed, predicted, probs, predictor = NULL)
}
\description{
create an \code{TwoWayClassification} instance
}

