% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mvpa_model.R
\name{TwoWayClassificationResult}
\alias{TwoWayClassificationResult}
\title{create an \code{TwoWayClassification} instance}
\usage{
TwoWayClassificationResult(observed, predicted, probs, predictor = NULL)
}
\arguments{
\item{observed}{}

\item{predicted}{}

\item{probs}{}
}
\description{
create an \code{TwoWayClassification} instance
}

