% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_rsa_model.R
\name{vector_rsa_model}
\alias{vector_rsa_model}
\title{Create a vectorized RSA model}
\usage{
vector_rsa_model(dataset, design)
}
\arguments{
\item{design}{A vector_rsa_design object.}
}
\value{
An object representing the RSA model, which includes both the design and precomputed data.
}
\description{
This function integrates a vector_rsa_design and precomputes data to create a vectorized RSA model.
}
\details{
Integrates RSA design and precomputed data to facilitate efficient RSA computations. 
It directly incorporates precomputations for cross-block comparisons.
}
