% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_rsa_model.R
\name{vector_rsa_model}
\alias{vector_rsa_model}
\title{Create a vectorized RSA model}
\usage{
vector_rsa_model(
  dataset,
  design,
  distfun = cordist(),
  rsa_simfun = c("pearson", "spearman")
)
}
\arguments{
\item{dataset}{An \code{mvpa_dataset}  object.}

\item{design}{A vector_rsa_design object.}

\item{distfun}{an object of \code{distfun} type for computation of pairwise dissimilarities among image rows}

\item{rsa_simfun}{a character string specifying the similarity function to use for RSA}
}
\value{
An object representing the RSA model, which includes both the design and precomputed data.
}
\description{
This function integrates a vector_rsa_design and precomputes data to create a vectorized RSA model.
}
\details{
Integrates RSA design and precomputed data to facilitate efficient RSA computations. 
It directly incorporates precomputations for cross-block comparisons.
}
