% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{weighted_model}
\alias{weighted_model}
\title{weighted_model}
\usage{
weighted_model(fits, names = 1:length(fits), weights = rep(1/length(fits),
  length(fits)))
}
\arguments{
\item{fits}{a list of model fits}

\item{names}{a list of names, one per model fit}

\item{weights}{a vector of weights, one per model fit}
}
\description{
a consensus model formed as a weighted average of a set of models
}
