% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{wrap_output}
\alias{wrap_output}
\title{Wrap Output}
\usage{
wrap_output(obj, vals, ...)
}
\arguments{
\item{obj}{The object used to determine the wrapping method.}

\item{vals}{The values to be wrapped.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A wrapped output object.
}
\description{
Wrap output values into a desired format.
}
\keyword{internal}
