% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R
\name{wrap_output}
\alias{wrap_output}
\title{wrap_output}
\usage{
wrap_output(obj, vals, ...)
}
\arguments{
\item{obj}{the object}

\item{vals}{the values to wrap}

\item{...}{extra args}
}
\description{
wrap_output
}
