% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgeneric.R, R/design.R, R/item_design.R,
%   R/mvpa_model.R, R/feature_rsa_model.R, R/hrfdecoder_model.R
\name{y_train}
\alias{y_train}
\alias{y_train.mvpa_design}
\alias{y_train.item_design}
\alias{y_train.mvpa_model}
\alias{y_train.feature_rsa_model}
\alias{y_train.feature_rsa_design}
\alias{y_train.hrfdecoder_model}
\title{Training Labels/Response Extraction}
\usage{
y_train(obj)

\method{y_train}{mvpa_design}(obj)

\method{y_train}{item_design}(obj)

\method{y_train}{mvpa_model}(obj)

\method{y_train}{feature_rsa_model}(obj)

\method{y_train}{feature_rsa_design}(obj)

\method{y_train}{hrfdecoder_model}(obj)
}
\arguments{
\item{obj}{The object from which to extract the training response variable.}
}
\value{
The training response variable (factor for classification, numeric for regression).
}
\description{
Extract the training labels or response variable from an object.
}
\examples{
ds <- gen_sample_dataset(D = c(4, 4, 4), nobs = 10)
y_train(ds$design)
}
